\name{HSSVD-package}
\alias{HSSVD-package}
\alias{HSSVD}
\docType{package}
\title{
Biclustering with heterogeneous variance
}
\description{
HSSVD is a recently developed data mining tool for discovering subgroups of 
patients and genes which simultaneously display unusual levels of variability 
compared to other genes and patients. Previous biclustering methods were 
restricted to mean level detection, while the new method can detect both 
mean and variance biclusters.}
\details{
\tabular{ll}{
Package: \tab HSSVD\cr
Type: \tab Package\cr
Version: \tab 1.1\cr
Date: \tab 2014-07-21\cr
License: \tab GPL-2\cr
}
low_rank is the full implementation of the HSSVD framework.
}
\author{
Authors: Guanhua Chen and Michael R. Kosorok. \cr
Contributor: Shannon T. Holloway <sthollow@ncsu.edu> \cr
Maintainer: Guanhua Chen <guanhuac@live.unc.edu>
}
\references{
Chen G, Sullivan PF, Kosorok MR. "Biclustering with heterogeneous variance." 
Proc. Natl. Acad. Sci. U.S.A.. 2013;110(30):12253-8.

Owen AB, Perry PO (2009) Bi-cross-validation of the SVD and the nonnegative matrix
factorization. The Annals of Applied Statistics 3:564-594.

Yang D, Zongming M, Buja, A. "A Sparse SVD Method for High-dimensional Data."
arXiv:11112.24333 (2011).}
\keyword{ package }
