\name{Methylation}
\alias{Methylation}
\docType{data}
\title{
Methylation Data in cancer versus normal patients
}
\description{
Description
}
\usage{data(Methylation)}
\format{
  A data frame with 384 observations on the following 1465 variables.
  \describe{
    \item{\code{TargetID}}{a factor with levels }
    \item{\code{CHROMOSOME}}{a factor with levels }
    \item{\code{SYMBOL}}{a factor with levels }
    \item{\code{CPG_ISLAND}}{a logical vector}
    \item{\code{PRODUCT}}{a logical vector}
    \item{\code{ANNOTATION}}{a logical vector}
    \item{\code{SYNONYM}}{a logical vector}
    \item{\code{INPUT_SEQUENCE}}{a factor with levels }
    \item{\code{REFSEQ}}{a numeric vector}
    \item{\code{CG_NO}}{a factor with levels }
    \item{\code{DIST_TO_TSS}}{a logical vector}
    \item{\code{CPG_COORDINATE}}{a numeric vector}
    \item{\code{GID}}{a logical vector}
    \item{\code{ProbeID}}{a numeric vector}
    \item{\code{GENE_ID}}{a logical vector}
    \item{\code{Breast_N1.1034270290.AVG_Beta}}{a numeric vector}
    \item{\code{Breast_N1.1034270290.Signal.CY3}}{a numeric vector}
    \item{\code{Breast_N1.1034270290.Signal.CY5}}{a numeric vector}
    \item{\code{Breast_N1.1034270290.Avg_NBEADS}}{a numeric vector}
    \item{\code{Breast_N1.1034270290.BEAD_STDERR}}{a numeric vector}
    \item{\code{Breast_N10.1034269692.AVG_Beta}}{a numeric vector}
    \item{\code{Breast_N10.1034269692.Signal.CY3}}{a numeric vector}
    \item{\code{Breast_N10.1034269692.Signal.CY5}}{a numeric vector}
    \item{\code{Breast_N10.1034269692.Avg_NBEADS}}{a numeric vector}
    \item{\code{Breast_N10.1034269692.BEAD_STDERR}}{a numeric vector}
    \item{\code{Breast_N11.1034271163.AVG_Beta}}{a numeric vector}
    \item{\code{Breast_N11.1034271163.Signal.CY3}}{a numeric vector}
    \item{\code{Breast_N11.1034271163.Signal.CY5}}{a numeric vector}
    \item{\code{Breast_N11.1034271163.Avg_NBEADS}}{a numeric vector}
    \item{\code{Breast_N11.1034271163.BEAD_STDERR}}{a numeric vector}
    \item{\code{Breast_N12.1034271179.AVG_Beta}}{a numeric vector}
    \item{\code{Breast_N12.1034271179.Signal.CY3}}{a numeric vector}
    \item{\code{Breast_N12.1034271179.Signal.CY5}}{a numeric vector}
    \item{\code{Breast_N12.1034271179.Avg_NBEADS}}{a numeric vector}
    \item{\code{Breast_N12.1034271179.BEAD_STDERR}}{a numeric vector}
    \item{\code{Breast_N13.1034271188.AVG_Beta}}{a numeric vector}
    \item{\code{Breast_N13.1034271188.Signal.CY3}}{a numeric vector}
    \item{\code{Breast_N13.1034271188.Signal.CY5}}{a numeric vector}
    \item{\code{Breast_N13.1034271188.Avg_NBEADS}}{a numeric vector}
    \item{\code{Breast_N13.1034271188.BEAD_STDERR}}{a numeric vector}
    \item{\code{Breast_N14.1034271202.AVG_Beta}}{a numeric vector}
    \item{\code{Breast_N14.1034271202.Signal.CY3}}{a numeric vector}
    \item{\code{Breast_N14.1034271202.Signal.CY5}}{a numeric vector}
    \item{\code{Breast_N14.1034271202.Avg_NBEADS}}{a numeric vector}
    \item{\code{Breast_N14.1034271202.BEAD_STDERR}}{a numeric vector}
    \item{\code{Breast_N15.1034271214.AVG_Beta}}{a numeric vector}
    \item{\code{Breast_N15.1034271214.Signal.CY3}}{a numeric vector}
    \item{\code{Breast_N15.1034271214.Signal.CY5}}{a numeric vector}
    \item{\code{Breast_N15.1034271214.Avg_NBEADS}}{a numeric vector}
    \item{\code{Breast_N15.1034271214.BEAD_STDERR}}{a numeric vector}
    \item{\code{Breast_N16.1034271223.AVG_Beta}}{a numeric vector}
    \item{\code{Breast_N16.1034271223.Signal.CY3}}{a numeric vector}
    \item{\code{Breast_N16.1034271223.Signal.CY5}}{a numeric vector}
    \item{\code{Breast_N16.1034271223.Avg_NBEADS}}{a numeric vector}
    \item{\code{Breast_N16.1034271223.BEAD_STDERR}}{a numeric vector}
    \item{\code{Breast_N17.1034271241.AVG_Beta}}{a numeric vector}
    \item{\code{Breast_N17.1034271241.Signal.CY3}}{a numeric vector}
    \item{\code{Breast_N17.1034271241.Signal.CY5}}{a numeric vector}
    \item{\code{Breast_N17.1034271241.Avg_NBEADS}}{a numeric vector}
    \item{\code{Breast_N17.1034271241.BEAD_STDERR}}{a numeric vector}
    \item{\code{Breast_N18.1034247347.AVG_Beta}}{a numeric vector}
    \item{\code{Breast_N18.1034247347.Signal.CY3}}{a numeric vector}
    \item{\code{Breast_N18.1034247347.Signal.CY5}}{a numeric vector}
    \item{\code{Breast_N18.1034247347.Avg_NBEADS}}{a numeric vector}
    \item{\code{Breast_N18.1034247347.BEAD_STDERR}}{a numeric vector}
    \item{\code{Breast_N19.1034247365.AVG_Beta}}{a numeric vector}
    \item{\code{Breast_N19.1034247365.Signal.CY3}}{a numeric vector}
    \item{\code{Breast_N19.1034247365.Signal.CY5}}{a numeric vector}
    \item{\code{Breast_N19.1034247365.Avg_NBEADS}}{a numeric vector}
    \item{\code{Breast_N19.1034247365.BEAD_STDERR}}{a numeric vector}
    \item{\code{Breast_N2.1034270303.AVG_Beta}}{a numeric vector}
    \item{\code{Breast_N2.1034270303.Signal.CY3}}{a numeric vector}
    \item{\code{Breast_N2.1034270303.Signal.CY5}}{a numeric vector}
    \item{\code{Breast_N2.1034270303.Avg_NBEADS}}{a numeric vector}
    \item{\code{Breast_N2.1034270303.BEAD_STDERR}}{a numeric vector}
    \item{\code{Breast_N20.1034247368.AVG_Beta}}{a numeric vector}
    \item{\code{Breast_N20.1034247368.Signal.CY3}}{a numeric vector}
    \item{\code{Breast_N20.1034247368.Signal.CY5}}{a numeric vector}
    \item{\code{Breast_N20.1034247368.Avg_NBEADS}}{a numeric vector}
    \item{\code{Breast_N20.1034247368.BEAD_STDERR}}{a numeric vector}
    \item{\code{Breast_N21.1034247391.AVG_Beta}}{a numeric vector}
    \item{\code{Breast_N21.1034247391.Signal.CY3}}{a numeric vector}
    \item{\code{Breast_N21.1034247391.Signal.CY5}}{a numeric vector}
    \item{\code{Breast_N21.1034247391.Avg_NBEADS}}{a numeric vector}
    \item{\code{Breast_N21.1034247391.BEAD_STDERR}}{a numeric vector}
    \item{\code{Breast_N22.1034247380.AVG_Beta}}{a numeric vector}
    \item{\code{Breast_N22.1034247380.Signal.CY3}}{a numeric vector}
    \item{\code{Breast_N22.1034247380.Signal.CY5}}{a numeric vector}
    \item{\code{Breast_N22.1034247380.Avg_NBEADS}}{a numeric vector}
    \item{\code{Breast_N22.1034247380.BEAD_STDERR}}{a numeric vector}
    \item{\code{Breast_N23.1034247402.AVG_Beta}}{a numeric vector}
    \item{\code{Breast_N23.1034247402.Signal.CY3}}{a numeric vector}
    \item{\code{Breast_N23.1034247402.Signal.CY5}}{a numeric vector}
    \item{\code{Breast_N23.1034247402.Avg_NBEADS}}{a numeric vector}
    \item{\code{Breast_N23.1034247402.BEAD_STDERR}}{a numeric vector}
    \item{\code{Breast_N24.1034247409.AVG_Beta}}{a numeric vector}
    \item{\code{Breast_N24.1034247409.Signal.CY3}}{a numeric vector}
    \item{\code{Breast_N24.1034247409.Signal.CY5}}{a numeric vector}
    \item{\code{Breast_N24.1034247409.Avg_NBEADS}}{a numeric vector}
    \item{\code{Breast_N24.1034247409.BEAD_STDERR}}{a numeric vector}
    \item{\code{Breast_N25.1034247419.AVG_Beta}}{a numeric vector}
    \item{\code{Breast_N25.1034247419.Signal.CY3}}{a numeric vector}
    \item{\code{Breast_N25.1034247419.Signal.CY5}}{a numeric vector}
    \item{\code{Breast_N25.1034247419.Avg_NBEADS}}{a numeric vector}
    \item{\code{Breast_N25.1034247419.BEAD_STDERR}}{a numeric vector}
    \item{\code{Breast_N26.1034247426.AVG_Beta}}{a numeric vector}
    \item{\code{Breast_N26.1034247426.Signal.CY3}}{a numeric vector}
    \item{\code{Breast_N26.1034247426.Signal.CY5}}{a numeric vector}
    \item{\code{Breast_N26.1034247426.Avg_NBEADS}}{a numeric vector}
    \item{\code{Breast_N26.1034247426.BEAD_STDERR}}{a numeric vector}
    \item{\code{Breast_T3.1034270311.AVG_Beta}}{a numeric vector}
    \item{\code{Breast_T3.1034270311.Signal.CY3}}{a numeric vector}
    \item{\code{Breast_T3.1034270311.Signal.CY5}}{a numeric vector}
    \item{\code{Breast_T3.1034270311.Avg_NBEADS}}{a numeric vector}
    \item{\code{Breast_T3.1034270311.BEAD_STDERR}}{a numeric vector}
    \item{\code{Breast_T4.1034270328.AVG_Beta}}{a numeric vector}
    \item{\code{Breast_T4.1034270328.Signal.CY3}}{a numeric vector}
    \item{\code{Breast_T4.1034270328.Signal.CY5}}{a numeric vector}
    \item{\code{Breast_T4.1034270328.Avg_NBEADS}}{a numeric vector}
    \item{\code{Breast_T4.1034270328.BEAD_STDERR}}{a numeric vector}
    \item{\code{Breast_N5.1034270332.AVG_Beta}}{a numeric vector}
    \item{\code{Breast_N5.1034270332.Signal.CY3}}{a numeric vector}
    \item{\code{Breast_N5.1034270332.Signal.CY5}}{a numeric vector}
    \item{\code{Breast_N5.1034270332.Avg_NBEADS}}{a numeric vector}
    \item{\code{Breast_N5.1034270332.BEAD_STDERR}}{a numeric vector}
    \item{\code{Breast_N6.1034270355.AVG_Beta}}{a numeric vector}
    \item{\code{Breast_N6.1034270355.Signal.CY3}}{a numeric vector}
    \item{\code{Breast_N6.1034270355.Signal.CY5}}{a numeric vector}
    \item{\code{Breast_N6.1034270355.Avg_NBEADS}}{a numeric vector}
    \item{\code{Breast_N6.1034270355.BEAD_STDERR}}{a numeric vector}
    \item{\code{Breast_N7.1034270345.AVG_Beta}}{a numeric vector}
    \item{\code{Breast_N7.1034270345.Signal.CY3}}{a numeric vector}
    \item{\code{Breast_N7.1034270345.Signal.CY5}}{a numeric vector}
    \item{\code{Breast_N7.1034270345.Avg_NBEADS}}{a numeric vector}
    \item{\code{Breast_N7.1034270345.BEAD_STDERR}}{a numeric vector}
    \item{\code{Breast_N8.1034270366.AVG_Beta}}{a numeric vector}
    \item{\code{Breast_N8.1034270366.Signal.CY3}}{a numeric vector}
    \item{\code{Breast_N8.1034270366.Signal.CY5}}{a numeric vector}
    \item{\code{Breast_N8.1034270366.Avg_NBEADS}}{a numeric vector}
    \item{\code{Breast_N8.1034270366.BEAD_STDERR}}{a numeric vector}
    \item{\code{Breast_N9.1034270369.AVG_Beta}}{a numeric vector}
    \item{\code{Breast_N9.1034270369.Signal.CY3}}{a numeric vector}
    \item{\code{Breast_N9.1034270369.Signal.CY5}}{a numeric vector}
    \item{\code{Breast_N9.1034270369.Avg_NBEADS}}{a numeric vector}
    \item{\code{Breast_N9.1034270369.BEAD_STDERR}}{a numeric vector}
    \item{\code{Breast_T1.1034270289.AVG_Beta}}{a numeric vector}
    \item{\code{Breast_T1.1034270289.Signal.CY3}}{a numeric vector}
    \item{\code{Breast_T1.1034270289.Signal.CY5}}{a numeric vector}
    \item{\code{Breast_T1.1034270289.Avg_NBEADS}}{a numeric vector}
    \item{\code{Breast_T1.1034270289.BEAD_STDERR}}{a numeric vector}
    \item{\code{Breast_T10.1034271156.AVG_Beta}}{a numeric vector}
    \item{\code{Breast_T10.1034271156.Signal.CY3}}{a numeric vector}
    \item{\code{Breast_T10.1034271156.Signal.CY5}}{a numeric vector}
    \item{\code{Breast_T10.1034271156.Avg_NBEADS}}{a numeric vector}
    \item{\code{Breast_T10.1034271156.BEAD_STDERR}}{a numeric vector}
    \item{\code{Breast_T11.1034271164.AVG_Beta}}{a numeric vector}
    \item{\code{Breast_T11.1034271164.Signal.CY3}}{a numeric vector}
    \item{\code{Breast_T11.1034271164.Signal.CY5}}{a numeric vector}
    \item{\code{Breast_T11.1034271164.Avg_NBEADS}}{a numeric vector}
    \item{\code{Breast_T11.1034271164.BEAD_STDERR}}{a numeric vector}
    \item{\code{Breast_T12.1034271178.AVG_Beta}}{a numeric vector}
    \item{\code{Breast_T12.1034271178.Signal.CY3}}{a numeric vector}
    \item{\code{Breast_T12.1034271178.Signal.CY5}}{a numeric vector}
    \item{\code{Breast_T12.1034271178.Avg_NBEADS}}{a numeric vector}
    \item{\code{Breast_T12.1034271178.BEAD_STDERR}}{a numeric vector}
    \item{\code{Breast_T13.1034271190.AVG_Beta}}{a numeric vector}
    \item{\code{Breast_T13.1034271190.Signal.CY3}}{a numeric vector}
    \item{\code{Breast_T13.1034271190.Signal.CY5}}{a numeric vector}
    \item{\code{Breast_T13.1034271190.Avg_NBEADS}}{a numeric vector}
    \item{\code{Breast_T13.1034271190.BEAD_STDERR}}{a numeric vector}
    \item{\code{Breast_T14.1034271201.AVG_Beta}}{a numeric vector}
    \item{\code{Breast_T14.1034271201.Signal.CY3}}{a numeric vector}
    \item{\code{Breast_T14.1034271201.Signal.CY5}}{a numeric vector}
    \item{\code{Breast_T14.1034271201.Avg_NBEADS}}{a numeric vector}
    \item{\code{Breast_T14.1034271201.BEAD_STDERR}}{a numeric vector}
    \item{\code{Breast_T15.1034271215.AVG_Beta}}{a numeric vector}
    \item{\code{Breast_T15.1034271215.Signal.CY3}}{a numeric vector}
    \item{\code{Breast_T15.1034271215.Signal.CY5}}{a numeric vector}
    \item{\code{Breast_T15.1034271215.Avg_NBEADS}}{a numeric vector}
    \item{\code{Breast_T15.1034271215.BEAD_STDERR}}{a numeric vector}
    \item{\code{Breast_T16.1034271222.AVG_Beta}}{a numeric vector}
    \item{\code{Breast_T16.1034271222.Signal.CY3}}{a numeric vector}
    \item{\code{Breast_T16.1034271222.Signal.CY5}}{a numeric vector}
    \item{\code{Breast_T16.1034271222.Avg_NBEADS}}{a numeric vector}
    \item{\code{Breast_T16.1034271222.BEAD_STDERR}}{a numeric vector}
    \item{\code{Breast_T17.1034271242.AVG_Beta}}{a numeric vector}
    \item{\code{Breast_T17.1034271242.Signal.CY3}}{a numeric vector}
    \item{\code{Breast_T17.1034271242.Signal.CY5}}{a numeric vector}
    \item{\code{Breast_T17.1034271242.Avg_NBEADS}}{a numeric vector}
    \item{\code{Breast_T17.1034271242.BEAD_STDERR}}{a numeric vector}
    \item{\code{Breast_T18.1034247346.AVG_Beta}}{a numeric vector}
    \item{\code{Breast_T18.1034247346.Signal.CY3}}{a numeric vector}
    \item{\code{Breast_T18.1034247346.Signal.CY5}}{a numeric vector}
    \item{\code{Breast_T18.1034247346.Avg_NBEADS}}{a numeric vector}
    \item{\code{Breast_T18.1034247346.BEAD_STDERR}}{a numeric vector}
    \item{\code{Breast_T19.1034247367.AVG_Beta}}{a numeric vector}
    \item{\code{Breast_T19.1034247367.Signal.CY3}}{a numeric vector}
    \item{\code{Breast_T19.1034247367.Signal.CY5}}{a numeric vector}
    \item{\code{Breast_T19.1034247367.Avg_NBEADS}}{a numeric vector}
    \item{\code{Breast_T19.1034247367.BEAD_STDERR}}{a numeric vector}
    \item{\code{Breast_T2.1034270304.AVG_Beta}}{a numeric vector}
    \item{\code{Breast_T2.1034270304.Signal.CY3}}{a numeric vector}
    \item{\code{Breast_T2.1034270304.Signal.CY5}}{a numeric vector}
    \item{\code{Breast_T2.1034270304.Avg_NBEADS}}{a numeric vector}
    \item{\code{Breast_T2.1034270304.BEAD_STDERR}}{a numeric vector}
    \item{\code{Breast_T20.1034247369.AVG_Beta}}{a numeric vector}
    \item{\code{Breast_T20.1034247369.Signal.CY3}}{a numeric vector}
    \item{\code{Breast_T20.1034247369.Signal.CY5}}{a numeric vector}
    \item{\code{Breast_T20.1034247369.Avg_NBEADS}}{a numeric vector}
    \item{\code{Breast_T20.1034247369.BEAD_STDERR}}{a numeric vector}
    \item{\code{Breast_T21.1034247379.AVG_Beta}}{a numeric vector}
    \item{\code{Breast_T21.1034247379.Signal.CY3}}{a numeric vector}
    \item{\code{Breast_T21.1034247379.Signal.CY5}}{a numeric vector}
    \item{\code{Breast_T21.1034247379.Avg_NBEADS}}{a numeric vector}
    \item{\code{Breast_T21.1034247379.BEAD_STDERR}}{a numeric vector}
    \item{\code{Breast_T22.1034247381.AVG_Beta}}{a numeric vector}
    \item{\code{Breast_T22.1034247381.Signal.CY3}}{a numeric vector}
    \item{\code{Breast_T22.1034247381.Signal.CY5}}{a numeric vector}
    \item{\code{Breast_T22.1034247381.Avg_NBEADS}}{a numeric vector}
    \item{\code{Breast_T22.1034247381.BEAD_STDERR}}{a numeric vector}
    \item{\code{Breast_T23.1034247401.AVG_Beta}}{a numeric vector}
    \item{\code{Breast_T23.1034247401.Signal.CY3}}{a numeric vector}
    \item{\code{Breast_T23.1034247401.Signal.CY5}}{a numeric vector}
    \item{\code{Breast_T23.1034247401.Avg_NBEADS}}{a numeric vector}
    \item{\code{Breast_T23.1034247401.BEAD_STDERR}}{a numeric vector}
    \item{\code{Breast_T24.1034247410.AVG_Beta}}{a numeric vector}
    \item{\code{Breast_T24.1034247410.Signal.CY3}}{a numeric vector}
    \item{\code{Breast_T24.1034247410.Signal.CY5}}{a numeric vector}
    \item{\code{Breast_T24.1034247410.Avg_NBEADS}}{a numeric vector}
    \item{\code{Breast_T24.1034247410.BEAD_STDERR}}{a numeric vector}
    \item{\code{Breast_T25.1034247418.AVG_Beta}}{a numeric vector}
    \item{\code{Breast_T25.1034247418.Signal.CY3}}{a numeric vector}
    \item{\code{Breast_T25.1034247418.Signal.CY5}}{a numeric vector}
    \item{\code{Breast_T25.1034247418.Avg_NBEADS}}{a numeric vector}
    \item{\code{Breast_T25.1034247418.BEAD_STDERR}}{a numeric vector}
    \item{\code{Breast_T26.1034247425.AVG_Beta}}{a numeric vector}
    \item{\code{Breast_T26.1034247425.Signal.CY3}}{a numeric vector}
    \item{\code{Breast_T26.1034247425.Signal.CY5}}{a numeric vector}
    \item{\code{Breast_T26.1034247425.Avg_NBEADS}}{a numeric vector}
    \item{\code{Breast_T26.1034247425.BEAD_STDERR}}{a numeric vector}
    \item{\code{Breast_N3.1034270310.AVG_Beta}}{a numeric vector}
    \item{\code{Breast_N3.1034270310.Signal.CY3}}{a numeric vector}
    \item{\code{Breast_N3.1034270310.Signal.CY5}}{a numeric vector}
    \item{\code{Breast_N3.1034270310.Avg_NBEADS}}{a numeric vector}
    \item{\code{Breast_N3.1034270310.BEAD_STDERR}}{a numeric vector}
    \item{\code{Breast_N4.1034270329.AVG_Beta}}{a numeric vector}
    \item{\code{Breast_N4.1034270329.Signal.CY3}}{a numeric vector}
    \item{\code{Breast_N4.1034270329.Signal.CY5}}{a numeric vector}
    \item{\code{Breast_N4.1034270329.Avg_NBEADS}}{a numeric vector}
    \item{\code{Breast_N4.1034270329.BEAD_STDERR}}{a numeric vector}
    \item{\code{Breast_T5.1034270333.AVG_Beta}}{a numeric vector}
    \item{\code{Breast_T5.1034270333.Signal.CY3}}{a numeric vector}
    \item{\code{Breast_T5.1034270333.Signal.CY5}}{a numeric vector}
    \item{\code{Breast_T5.1034270333.Avg_NBEADS}}{a numeric vector}
    \item{\code{Breast_T5.1034270333.BEAD_STDERR}}{a numeric vector}
    \item{\code{Breast_T6.1034269657.AVG_Beta}}{a numeric vector}
    \item{\code{Breast_T6.1034269657.Signal.CY3}}{a numeric vector}
    \item{\code{Breast_T6.1034269657.Signal.CY5}}{a numeric vector}
    \item{\code{Breast_T6.1034269657.Avg_NBEADS}}{a numeric vector}
    \item{\code{Breast_T6.1034269657.BEAD_STDERR}}{a numeric vector}
    \item{\code{Colon_N1.1034270292.AVG_Beta}}{a numeric vector}
    \item{\code{Colon_N1.1034270292.Signal.CY3}}{a numeric vector}
    \item{\code{Colon_N1.1034270292.Signal.CY5}}{a numeric vector}
    \item{\code{Colon_N1.1034270292.Avg_NBEADS}}{a numeric vector}
    \item{\code{Colon_N1.1034270292.BEAD_STDERR}}{a numeric vector}
    \item{\code{Colon_N10.1034269701.AVG_Beta}}{a numeric vector}
    \item{\code{Colon_N10.1034269701.Signal.CY3}}{a numeric vector}
    \item{\code{Colon_N10.1034269701.Signal.CY5}}{a numeric vector}
    \item{\code{Colon_N10.1034269701.Avg_NBEADS}}{a numeric vector}
    \item{\code{Colon_N10.1034269701.BEAD_STDERR}}{a numeric vector}
    \item{\code{Colon_N11.1034271160.AVG_Beta}}{a numeric vector}
    \item{\code{Colon_N11.1034271160.Signal.CY3}}{a numeric vector}
    \item{\code{Colon_N11.1034271160.Signal.CY5}}{a numeric vector}
    \item{\code{Colon_N11.1034271160.Avg_NBEADS}}{a numeric vector}
    \item{\code{Colon_N11.1034271160.BEAD_STDERR}}{a numeric vector}
    \item{\code{Colon_N12.1034271181.AVG_Beta}}{a numeric vector}
    \item{\code{Colon_N12.1034271181.Signal.CY3}}{a numeric vector}
    \item{\code{Colon_N12.1034271181.Signal.CY5}}{a numeric vector}
    \item{\code{Colon_N12.1034271181.Avg_NBEADS}}{a numeric vector}
    \item{\code{Colon_N12.1034271181.BEAD_STDERR}}{a numeric vector}
    \item{\code{Colon_N14.1034271204.AVG_Beta}}{a numeric vector}
    \item{\code{Colon_N14.1034271204.Signal.CY3}}{a numeric vector}
    \item{\code{Colon_N14.1034271204.Signal.CY5}}{a numeric vector}
    \item{\code{Colon_N14.1034271204.Avg_NBEADS}}{a numeric vector}
    \item{\code{Colon_N14.1034271204.BEAD_STDERR}}{a numeric vector}
    \item{\code{Colon_N15.1034271212.AVG_Beta}}{a numeric vector}
    \item{\code{Colon_N15.1034271212.Signal.CY3}}{a numeric vector}
    \item{\code{Colon_N15.1034271212.Signal.CY5}}{a numeric vector}
    \item{\code{Colon_N15.1034271212.Avg_NBEADS}}{a numeric vector}
    \item{\code{Colon_N15.1034271212.BEAD_STDERR}}{a numeric vector}
    \item{\code{Colon_N16.1034271225.AVG_Beta}}{a numeric vector}
    \item{\code{Colon_N16.1034271225.Signal.CY3}}{a numeric vector}
    \item{\code{Colon_N16.1034271225.Signal.CY5}}{a numeric vector}
    \item{\code{Colon_N16.1034271225.Avg_NBEADS}}{a numeric vector}
    \item{\code{Colon_N16.1034271225.BEAD_STDERR}}{a numeric vector}
    \item{\code{Colon_N17.1034271239.AVG_Beta}}{a numeric vector}
    \item{\code{Colon_N17.1034271239.Signal.CY3}}{a numeric vector}
    \item{\code{Colon_N17.1034271239.Signal.CY5}}{a numeric vector}
    \item{\code{Colon_N17.1034271239.Avg_NBEADS}}{a numeric vector}
    \item{\code{Colon_N17.1034271239.BEAD_STDERR}}{a numeric vector}
    \item{\code{Colon_N18.1034247349.AVG_Beta}}{a numeric vector}
    \item{\code{Colon_N18.1034247349.Signal.CY3}}{a numeric vector}
    \item{\code{Colon_N18.1034247349.Signal.CY5}}{a numeric vector}
    \item{\code{Colon_N18.1034247349.Avg_NBEADS}}{a numeric vector}
    \item{\code{Colon_N18.1034247349.BEAD_STDERR}}{a numeric vector}
    \item{\code{Colon_N19.1034247363.AVG_Beta}}{a numeric vector}
    \item{\code{Colon_N19.1034247363.Signal.CY3}}{a numeric vector}
    \item{\code{Colon_N19.1034247363.Signal.CY5}}{a numeric vector}
    \item{\code{Colon_N19.1034247363.Avg_NBEADS}}{a numeric vector}
    \item{\code{Colon_N19.1034247363.BEAD_STDERR}}{a numeric vector}
    \item{\code{Colon_N2.1034270301.AVG_Beta}}{a numeric vector}
    \item{\code{Colon_N2.1034270301.Signal.CY3}}{a numeric vector}
    \item{\code{Colon_N2.1034270301.Signal.CY5}}{a numeric vector}
    \item{\code{Colon_N2.1034270301.Avg_NBEADS}}{a numeric vector}
    \item{\code{Colon_N2.1034270301.BEAD_STDERR}}{a numeric vector}
    \item{\code{Colon_N20.1034247371.AVG_Beta}}{a numeric vector}
    \item{\code{Colon_N20.1034247371.Signal.CY3}}{a numeric vector}
    \item{\code{Colon_N20.1034247371.Signal.CY5}}{a numeric vector}
    \item{\code{Colon_N20.1034247371.Avg_NBEADS}}{a numeric vector}
    \item{\code{Colon_N20.1034247371.BEAD_STDERR}}{a numeric vector}
    \item{\code{Colon_N21.1034247389.AVG_Beta}}{a numeric vector}
    \item{\code{Colon_N21.1034247389.Signal.CY3}}{a numeric vector}
    \item{\code{Colon_N21.1034247389.Signal.CY5}}{a numeric vector}
    \item{\code{Colon_N21.1034247389.Avg_NBEADS}}{a numeric vector}
    \item{\code{Colon_N21.1034247389.BEAD_STDERR}}{a numeric vector}
    \item{\code{Colon_N22.1034247393.AVG_Beta}}{a numeric vector}
    \item{\code{Colon_N22.1034247393.Signal.CY3}}{a numeric vector}
    \item{\code{Colon_N22.1034247393.Signal.CY5}}{a numeric vector}
    \item{\code{Colon_N22.1034247393.Avg_NBEADS}}{a numeric vector}
    \item{\code{Colon_N22.1034247393.BEAD_STDERR}}{a numeric vector}
    \item{\code{Colon_N23.1034247415.AVG_Beta}}{a numeric vector}
    \item{\code{Colon_N23.1034247415.Signal.CY3}}{a numeric vector}
    \item{\code{Colon_N23.1034247415.Signal.CY5}}{a numeric vector}
    \item{\code{Colon_N23.1034247415.Avg_NBEADS}}{a numeric vector}
    \item{\code{Colon_N23.1034247415.BEAD_STDERR}}{a numeric vector}
    \item{\code{Colon_N24.1034247407.AVG_Beta}}{a numeric vector}
    \item{\code{Colon_N24.1034247407.Signal.CY3}}{a numeric vector}
    \item{\code{Colon_N24.1034247407.Signal.CY5}}{a numeric vector}
    \item{\code{Colon_N24.1034247407.Avg_NBEADS}}{a numeric vector}
    \item{\code{Colon_N24.1034247407.BEAD_STDERR}}{a numeric vector}
    \item{\code{Colon_N25.1034247421.AVG_Beta}}{a numeric vector}
    \item{\code{Colon_N25.1034247421.Signal.CY3}}{a numeric vector}
    \item{\code{Colon_N25.1034247421.Signal.CY5}}{a numeric vector}
    \item{\code{Colon_N25.1034247421.Avg_NBEADS}}{a numeric vector}
    \item{\code{Colon_N25.1034247421.BEAD_STDERR}}{a numeric vector}
    \item{\code{Colon_N26.1034247438.AVG_Beta}}{a numeric vector}
    \item{\code{Colon_N26.1034247438.Signal.CY3}}{a numeric vector}
    \item{\code{Colon_N26.1034247438.Signal.CY5}}{a numeric vector}
    \item{\code{Colon_N26.1034247438.Avg_NBEADS}}{a numeric vector}
    \item{\code{Colon_N26.1034247438.BEAD_STDERR}}{a numeric vector}
    \item{\code{Colon_N27.1034247434.AVG_Beta}}{a numeric vector}
    \item{\code{Colon_N27.1034247434.Signal.CY3}}{a numeric vector}
    \item{\code{Colon_N27.1034247434.Signal.CY5}}{a numeric vector}
    \item{\code{Colon_N27.1034247434.Avg_NBEADS}}{a numeric vector}
    \item{\code{Colon_N27.1034247434.BEAD_STDERR}}{a numeric vector}
    \item{\code{Colon_N3.1034270314.AVG_Beta}}{a numeric vector}
    \item{\code{Colon_N3.1034270314.Signal.CY3}}{a numeric vector}
    \item{\code{Colon_N3.1034270314.Signal.CY5}}{a numeric vector}
    \item{\code{Colon_N3.1034270314.Avg_NBEADS}}{a numeric vector}
    \item{\code{Colon_N3.1034270314.BEAD_STDERR}}{a numeric vector}
    \item{\code{Colon_N4.1034270326.AVG_Beta}}{a numeric vector}
    \item{\code{Colon_N4.1034270326.Signal.CY3}}{a numeric vector}
    \item{\code{Colon_N4.1034270326.Signal.CY5}}{a numeric vector}
    \item{\code{Colon_N4.1034270326.Avg_NBEADS}}{a numeric vector}
    \item{\code{Colon_N4.1034270326.BEAD_STDERR}}{a numeric vector}
    \item{\code{Colon_N5.1034270336.AVG_Beta}}{a numeric vector}
    \item{\code{Colon_N5.1034270336.Signal.CY3}}{a numeric vector}
    \item{\code{Colon_N5.1034270336.Signal.CY5}}{a numeric vector}
    \item{\code{Colon_N5.1034270336.Avg_NBEADS}}{a numeric vector}
    \item{\code{Colon_N5.1034270336.BEAD_STDERR}}{a numeric vector}
    \item{\code{Colon_N6.1034270353.AVG_Beta}}{a numeric vector}
    \item{\code{Colon_N6.1034270353.Signal.CY3}}{a numeric vector}
    \item{\code{Colon_N6.1034270353.Signal.CY5}}{a numeric vector}
    \item{\code{Colon_N6.1034270353.Avg_NBEADS}}{a numeric vector}
    \item{\code{Colon_N6.1034270353.BEAD_STDERR}}{a numeric vector}
    \item{\code{Colon_N7.1034270356.AVG_Beta}}{a numeric vector}
    \item{\code{Colon_N7.1034270356.Signal.CY3}}{a numeric vector}
    \item{\code{Colon_N7.1034270356.Signal.CY5}}{a numeric vector}
    \item{\code{Colon_N7.1034270356.Avg_NBEADS}}{a numeric vector}
    \item{\code{Colon_N7.1034270356.BEAD_STDERR}}{a numeric vector}
    \item{\code{Colon_N8.1034270378.AVG_Beta}}{a numeric vector}
    \item{\code{Colon_N8.1034270378.Signal.CY3}}{a numeric vector}
    \item{\code{Colon_N8.1034270378.Signal.CY5}}{a numeric vector}
    \item{\code{Colon_N8.1034270378.Avg_NBEADS}}{a numeric vector}
    \item{\code{Colon_N8.1034270378.BEAD_STDERR}}{a numeric vector}
    \item{\code{Colon_N9.1034269699.AVG_Beta}}{a numeric vector}
    \item{\code{Colon_N9.1034269699.Signal.CY3}}{a numeric vector}
    \item{\code{Colon_N9.1034269699.Signal.CY5}}{a numeric vector}
    \item{\code{Colon_N9.1034269699.Avg_NBEADS}}{a numeric vector}
    \item{\code{Colon_N9.1034269699.BEAD_STDERR}}{a numeric vector}
    \item{\code{Colon_T1.1034270291.AVG_Beta}}{a numeric vector}
    \item{\code{Colon_T1.1034270291.Signal.CY3}}{a numeric vector}
    \item{\code{Colon_T1.1034270291.Signal.CY5}}{a numeric vector}
    \item{\code{Colon_T1.1034270291.Avg_NBEADS}}{a numeric vector}
    \item{\code{Colon_T1.1034270291.BEAD_STDERR}}{a numeric vector}
    \item{\code{Colon_T10.1034269693.AVG_Beta}}{a numeric vector}
    \item{\code{Colon_T10.1034269693.Signal.CY3}}{a numeric vector}
    \item{\code{Colon_T10.1034269693.Signal.CY5}}{a numeric vector}
    \item{\code{Colon_T10.1034269693.Avg_NBEADS}}{a numeric vector}
    \item{\code{Colon_T10.1034269693.BEAD_STDERR}}{a numeric vector}
    \item{\code{Colon_T11.1034271162.AVG_Beta}}{a numeric vector}
    \item{\code{Colon_T11.1034271162.Signal.CY3}}{a numeric vector}
    \item{\code{Colon_T11.1034271162.Signal.CY5}}{a numeric vector}
    \item{\code{Colon_T11.1034271162.Avg_NBEADS}}{a numeric vector}
    \item{\code{Colon_T11.1034271162.BEAD_STDERR}}{a numeric vector}
    \item{\code{Colon_T13.1034271187.AVG_Beta}}{a numeric vector}
    \item{\code{Colon_T13.1034271187.Signal.CY3}}{a numeric vector}
    \item{\code{Colon_T13.1034271187.Signal.CY5}}{a numeric vector}
    \item{\code{Colon_T13.1034271187.Avg_NBEADS}}{a numeric vector}
    \item{\code{Colon_T13.1034271187.BEAD_STDERR}}{a numeric vector}
    \item{\code{Colon_T14.1034271203.AVG_Beta}}{a numeric vector}
    \item{\code{Colon_T14.1034271203.Signal.CY3}}{a numeric vector}
    \item{\code{Colon_T14.1034271203.Signal.CY5}}{a numeric vector}
    \item{\code{Colon_T14.1034271203.Avg_NBEADS}}{a numeric vector}
    \item{\code{Colon_T14.1034271203.BEAD_STDERR}}{a numeric vector}
    \item{\code{Colon_T17.1034271240.AVG_Beta}}{a numeric vector}
    \item{\code{Colon_T17.1034271240.Signal.CY3}}{a numeric vector}
    \item{\code{Colon_T17.1034271240.Signal.CY5}}{a numeric vector}
    \item{\code{Colon_T17.1034271240.Avg_NBEADS}}{a numeric vector}
    \item{\code{Colon_T17.1034271240.BEAD_STDERR}}{a numeric vector}
    \item{\code{Colon_T18.1034247348.AVG_Beta}}{a numeric vector}
    \item{\code{Colon_T18.1034247348.Signal.CY3}}{a numeric vector}
    \item{\code{Colon_T18.1034247348.Signal.CY5}}{a numeric vector}
    \item{\code{Colon_T18.1034247348.Avg_NBEADS}}{a numeric vector}
    \item{\code{Colon_T18.1034247348.BEAD_STDERR}}{a numeric vector}
    \item{\code{Colon_T19.1034247364.AVG_Beta}}{a numeric vector}
    \item{\code{Colon_T19.1034247364.Signal.CY3}}{a numeric vector}
    \item{\code{Colon_T19.1034247364.Signal.CY5}}{a numeric vector}
    \item{\code{Colon_T19.1034247364.Avg_NBEADS}}{a numeric vector}
    \item{\code{Colon_T19.1034247364.BEAD_STDERR}}{a numeric vector}
    \item{\code{Colon_T21.1034247390.AVG_Beta}}{a numeric vector}
    \item{\code{Colon_T21.1034247390.Signal.CY3}}{a numeric vector}
    \item{\code{Colon_T21.1034247390.Signal.CY5}}{a numeric vector}
    \item{\code{Colon_T21.1034247390.Avg_NBEADS}}{a numeric vector}
    \item{\code{Colon_T21.1034247390.BEAD_STDERR}}{a numeric vector}
    \item{\code{Colon_T22.1034247392.AVG_Beta}}{a numeric vector}
    \item{\code{Colon_T22.1034247392.Signal.CY3}}{a numeric vector}
    \item{\code{Colon_T22.1034247392.Signal.CY5}}{a numeric vector}
    \item{\code{Colon_T22.1034247392.Avg_NBEADS}}{a numeric vector}
    \item{\code{Colon_T22.1034247392.BEAD_STDERR}}{a numeric vector}
    \item{\code{Colon_T23.1034247403.AVG_Beta}}{a numeric vector}
    \item{\code{Colon_T23.1034247403.Signal.CY3}}{a numeric vector}
    \item{\code{Colon_T23.1034247403.Signal.CY5}}{a numeric vector}
    \item{\code{Colon_T23.1034247403.Avg_NBEADS}}{a numeric vector}
    \item{\code{Colon_T23.1034247403.BEAD_STDERR}}{a numeric vector}
    \item{\code{Colon_T24.1034247408.AVG_Beta}}{a numeric vector}
    \item{\code{Colon_T24.1034247408.Signal.CY3}}{a numeric vector}
    \item{\code{Colon_T24.1034247408.Signal.CY5}}{a numeric vector}
    \item{\code{Colon_T24.1034247408.Avg_NBEADS}}{a numeric vector}
    \item{\code{Colon_T24.1034247408.BEAD_STDERR}}{a numeric vector}
    \item{\code{Colon_T25.1034247420.AVG_Beta}}{a numeric vector}
    \item{\code{Colon_T25.1034247420.Signal.CY3}}{a numeric vector}
    \item{\code{Colon_T25.1034247420.Signal.CY5}}{a numeric vector}
    \item{\code{Colon_T25.1034247420.Avg_NBEADS}}{a numeric vector}
    \item{\code{Colon_T25.1034247420.BEAD_STDERR}}{a numeric vector}
    \item{\code{Colon_T27.1034247435.AVG_Beta}}{a numeric vector}
    \item{\code{Colon_T27.1034247435.Signal.CY3}}{a numeric vector}
    \item{\code{Colon_T27.1034247435.Signal.CY5}}{a numeric vector}
    \item{\code{Colon_T27.1034247435.Avg_NBEADS}}{a numeric vector}
    \item{\code{Colon_T27.1034247435.BEAD_STDERR}}{a numeric vector}
    \item{\code{Colon_T3.1034270313.AVG_Beta}}{a numeric vector}
    \item{\code{Colon_T3.1034270313.Signal.CY3}}{a numeric vector}
    \item{\code{Colon_T3.1034270313.Signal.CY5}}{a numeric vector}
    \item{\code{Colon_T3.1034270313.Avg_NBEADS}}{a numeric vector}
    \item{\code{Colon_T3.1034270313.BEAD_STDERR}}{a numeric vector}
    \item{\code{Colon_T4.1034270327.AVG_Beta}}{a numeric vector}
    \item{\code{Colon_T4.1034270327.Signal.CY3}}{a numeric vector}
    \item{\code{Colon_T4.1034270327.Signal.CY5}}{a numeric vector}
    \item{\code{Colon_T4.1034270327.Avg_NBEADS}}{a numeric vector}
    \item{\code{Colon_T4.1034270327.BEAD_STDERR}}{a numeric vector}
    \item{\code{Colon_T5.1034270335.AVG_Beta}}{a numeric vector}
    \item{\code{Colon_T5.1034270335.Signal.CY3}}{a numeric vector}
    \item{\code{Colon_T5.1034270335.Signal.CY5}}{a numeric vector}
    \item{\code{Colon_T5.1034270335.Avg_NBEADS}}{a numeric vector}
    \item{\code{Colon_T5.1034270335.BEAD_STDERR}}{a numeric vector}
    \item{\code{Colon_T6.1034270354.AVG_Beta}}{a numeric vector}
    \item{\code{Colon_T6.1034270354.Signal.CY3}}{a numeric vector}
    \item{\code{Colon_T6.1034270354.Signal.CY5}}{a numeric vector}
    \item{\code{Colon_T6.1034270354.Avg_NBEADS}}{a numeric vector}
    \item{\code{Colon_T6.1034270354.BEAD_STDERR}}{a numeric vector}
    \item{\code{Colon_T7.1034270357.AVG_Beta}}{a numeric vector}
    \item{\code{Colon_T7.1034270357.Signal.CY3}}{a numeric vector}
    \item{\code{Colon_T7.1034270357.Signal.CY5}}{a numeric vector}
    \item{\code{Colon_T7.1034270357.Avg_NBEADS}}{a numeric vector}
    \item{\code{Colon_T7.1034270357.BEAD_STDERR}}{a numeric vector}
    \item{\code{Colon_T8.1034270367.AVG_Beta}}{a numeric vector}
    \item{\code{Colon_T8.1034270367.Signal.CY3}}{a numeric vector}
    \item{\code{Colon_T8.1034270367.Signal.CY5}}{a numeric vector}
    \item{\code{Colon_T8.1034270367.Avg_NBEADS}}{a numeric vector}
    \item{\code{Colon_T8.1034270367.BEAD_STDERR}}{a numeric vector}
    \item{\code{Lung_N1.1034270288.AVG_Beta}}{a numeric vector}
    \item{\code{Lung_N1.1034270288.Signal.CY3}}{a numeric vector}
    \item{\code{Lung_N1.1034270288.Signal.CY5}}{a numeric vector}
    \item{\code{Lung_N1.1034270288.Avg_NBEADS}}{a numeric vector}
    \item{\code{Lung_N1.1034270288.BEAD_STDERR}}{a numeric vector}
    \item{\code{Lung_N10.1034271155.AVG_Beta}}{a numeric vector}
    \item{\code{Lung_N10.1034271155.Signal.CY3}}{a numeric vector}
    \item{\code{Lung_N10.1034271155.Signal.CY5}}{a numeric vector}
    \item{\code{Lung_N10.1034271155.Avg_NBEADS}}{a numeric vector}
    \item{\code{Lung_N10.1034271155.BEAD_STDERR}}{a numeric vector}
    \item{\code{Lung_N11.1034271165.AVG_Beta}}{a numeric vector}
    \item{\code{Lung_N11.1034271165.Signal.CY3}}{a numeric vector}
    \item{\code{Lung_N11.1034271165.Signal.CY5}}{a numeric vector}
    \item{\code{Lung_N11.1034271165.Avg_NBEADS}}{a numeric vector}
    \item{\code{Lung_N11.1034271165.BEAD_STDERR}}{a numeric vector}
    \item{\code{Lung_N12.1034271177.AVG_Beta}}{a numeric vector}
    \item{\code{Lung_N12.1034271177.Signal.CY3}}{a numeric vector}
    \item{\code{Lung_N12.1034271177.Signal.CY5}}{a numeric vector}
    \item{\code{Lung_N12.1034271177.Avg_NBEADS}}{a numeric vector}
    \item{\code{Lung_N12.1034271177.BEAD_STDERR}}{a numeric vector}
    \item{\code{Lung_N14.1034271200.AVG_Beta}}{a numeric vector}
    \item{\code{Lung_N14.1034271200.Signal.CY3}}{a numeric vector}
    \item{\code{Lung_N14.1034271200.Signal.CY5}}{a numeric vector}
    \item{\code{Lung_N14.1034271200.Avg_NBEADS}}{a numeric vector}
    \item{\code{Lung_N14.1034271200.BEAD_STDERR}}{a numeric vector}
    \item{\code{Lung_T15.1034271217.AVG_Beta}}{a numeric vector}
    \item{\code{Lung_T15.1034271217.Signal.CY3}}{a numeric vector}
    \item{\code{Lung_T15.1034271217.Signal.CY5}}{a numeric vector}
    \item{\code{Lung_T15.1034271217.Avg_NBEADS}}{a numeric vector}
    \item{\code{Lung_T15.1034271217.BEAD_STDERR}}{a numeric vector}
    \item{\code{Lung_T16.1034271220.AVG_Beta}}{a numeric vector}
    \item{\code{Lung_T16.1034271220.Signal.CY3}}{a numeric vector}
    \item{\code{Lung_T16.1034271220.Signal.CY5}}{a numeric vector}
    \item{\code{Lung_T16.1034271220.Avg_NBEADS}}{a numeric vector}
    \item{\code{Lung_T16.1034271220.BEAD_STDERR}}{a numeric vector}
    \item{\code{Lung_T17.1034271243.AVG_Beta}}{a numeric vector}
    \item{\code{Lung_T17.1034271243.Signal.CY3}}{a numeric vector}
    \item{\code{Lung_T17.1034271243.Signal.CY5}}{a numeric vector}
    \item{\code{Lung_T17.1034271243.Avg_NBEADS}}{a numeric vector}
    \item{\code{Lung_T17.1034271243.BEAD_STDERR}}{a numeric vector}
    \item{\code{Lung_T18.1034247345.AVG_Beta}}{a numeric vector}
    \item{\code{Lung_T18.1034247345.Signal.CY3}}{a numeric vector}
    \item{\code{Lung_T18.1034247345.Signal.CY5}}{a numeric vector}
    \item{\code{Lung_T18.1034247345.Avg_NBEADS}}{a numeric vector}
    \item{\code{Lung_T18.1034247345.BEAD_STDERR}}{a numeric vector}
    \item{\code{Lung_N19.1034247355.AVG_Beta}}{a numeric vector}
    \item{\code{Lung_N19.1034247355.Signal.CY3}}{a numeric vector}
    \item{\code{Lung_N19.1034247355.Signal.CY5}}{a numeric vector}
    \item{\code{Lung_N19.1034247355.Avg_NBEADS}}{a numeric vector}
    \item{\code{Lung_N19.1034247355.BEAD_STDERR}}{a numeric vector}
    \item{\code{Lung_N2.1034270305.AVG_Beta}}{a numeric vector}
    \item{\code{Lung_N2.1034270305.Signal.CY3}}{a numeric vector}
    \item{\code{Lung_N2.1034270305.Signal.CY5}}{a numeric vector}
    \item{\code{Lung_N2.1034270305.Avg_NBEADS}}{a numeric vector}
    \item{\code{Lung_N2.1034270305.BEAD_STDERR}}{a numeric vector}
    \item{\code{Lung_N20.1034247356.AVG_Beta}}{a numeric vector}
    \item{\code{Lung_N20.1034247356.Signal.CY3}}{a numeric vector}
    \item{\code{Lung_N20.1034247356.Signal.CY5}}{a numeric vector}
    \item{\code{Lung_N20.1034247356.Avg_NBEADS}}{a numeric vector}
    \item{\code{Lung_N20.1034247356.BEAD_STDERR}}{a numeric vector}
    \item{\code{Lung_N21.1034247377.AVG_Beta}}{a numeric vector}
    \item{\code{Lung_N21.1034247377.Signal.CY3}}{a numeric vector}
    \item{\code{Lung_N21.1034247377.Signal.CY5}}{a numeric vector}
    \item{\code{Lung_N21.1034247377.Avg_NBEADS}}{a numeric vector}
    \item{\code{Lung_N21.1034247377.BEAD_STDERR}}{a numeric vector}
    \item{\code{Lung_N22.1034247382.AVG_Beta}}{a numeric vector}
    \item{\code{Lung_N22.1034247382.Signal.CY3}}{a numeric vector}
    \item{\code{Lung_N22.1034247382.Signal.CY5}}{a numeric vector}
    \item{\code{Lung_N22.1034247382.Avg_NBEADS}}{a numeric vector}
    \item{\code{Lung_N22.1034247382.BEAD_STDERR}}{a numeric vector}
    \item{\code{Lung_N23.1034247400.AVG_Beta}}{a numeric vector}
    \item{\code{Lung_N23.1034247400.Signal.CY3}}{a numeric vector}
    \item{\code{Lung_N23.1034247400.Signal.CY5}}{a numeric vector}
    \item{\code{Lung_N23.1034247400.Avg_NBEADS}}{a numeric vector}
    \item{\code{Lung_N23.1034247400.BEAD_STDERR}}{a numeric vector}
    \item{\code{Lung_N24.1034247411.AVG_Beta}}{a numeric vector}
    \item{\code{Lung_N24.1034247411.Signal.CY3}}{a numeric vector}
    \item{\code{Lung_N24.1034247411.Signal.CY5}}{a numeric vector}
    \item{\code{Lung_N24.1034247411.Avg_NBEADS}}{a numeric vector}
    \item{\code{Lung_N24.1034247411.BEAD_STDERR}}{a numeric vector}
    \item{\code{Lung_N25.1034247416.AVG_Beta}}{a numeric vector}
    \item{\code{Lung_N25.1034247416.Signal.CY3}}{a numeric vector}
    \item{\code{Lung_N25.1034247416.Signal.CY5}}{a numeric vector}
    \item{\code{Lung_N25.1034247416.Avg_NBEADS}}{a numeric vector}
    \item{\code{Lung_N25.1034247416.BEAD_STDERR}}{a numeric vector}
    \item{\code{Lung_N3.1034270308.AVG_Beta}}{a numeric vector}
    \item{\code{Lung_N3.1034270308.Signal.CY3}}{a numeric vector}
    \item{\code{Lung_N3.1034270308.Signal.CY5}}{a numeric vector}
    \item{\code{Lung_N3.1034270308.Avg_NBEADS}}{a numeric vector}
    \item{\code{Lung_N3.1034270308.BEAD_STDERR}}{a numeric vector}
    \item{\code{Lung_N4.1034270330.AVG_Beta}}{a numeric vector}
    \item{\code{Lung_N4.1034270330.Signal.CY3}}{a numeric vector}
    \item{\code{Lung_N4.1034270330.Signal.CY5}}{a numeric vector}
    \item{\code{Lung_N4.1034270330.Avg_NBEADS}}{a numeric vector}
    \item{\code{Lung_N4.1034270330.BEAD_STDERR}}{a numeric vector}
    \item{\code{Lung_N5.1034270320.AVG_Beta}}{a numeric vector}
    \item{\code{Lung_N5.1034270320.Signal.CY3}}{a numeric vector}
    \item{\code{Lung_N5.1034270320.Signal.CY5}}{a numeric vector}
    \item{\code{Lung_N5.1034270320.Avg_NBEADS}}{a numeric vector}
    \item{\code{Lung_N5.1034270320.BEAD_STDERR}}{a numeric vector}
    \item{\code{Lung_N7.1034270347.AVG_Beta}}{a numeric vector}
    \item{\code{Lung_N7.1034270347.Signal.CY3}}{a numeric vector}
    \item{\code{Lung_N7.1034270347.Signal.CY5}}{a numeric vector}
    \item{\code{Lung_N7.1034270347.Avg_NBEADS}}{a numeric vector}
    \item{\code{Lung_N7.1034270347.BEAD_STDERR}}{a numeric vector}
    \item{\code{Lung_N8.1034270364.AVG_Beta}}{a numeric vector}
    \item{\code{Lung_N8.1034270364.Signal.CY3}}{a numeric vector}
    \item{\code{Lung_N8.1034270364.Signal.CY5}}{a numeric vector}
    \item{\code{Lung_N8.1034270364.Avg_NBEADS}}{a numeric vector}
    \item{\code{Lung_N8.1034270364.BEAD_STDERR}}{a numeric vector}
    \item{\code{Lung_N9.1034270371.AVG_Beta}}{a numeric vector}
    \item{\code{Lung_N9.1034270371.Signal.CY3}}{a numeric vector}
    \item{\code{Lung_N9.1034270371.Signal.CY5}}{a numeric vector}
    \item{\code{Lung_N9.1034270371.Avg_NBEADS}}{a numeric vector}
    \item{\code{Lung_N9.1034270371.BEAD_STDERR}}{a numeric vector}
    \item{\code{Lung_T1.1034270287.AVG_Beta}}{a numeric vector}
    \item{\code{Lung_T1.1034270287.Signal.CY3}}{a numeric vector}
    \item{\code{Lung_T1.1034270287.Signal.CY5}}{a numeric vector}
    \item{\code{Lung_T1.1034270287.Avg_NBEADS}}{a numeric vector}
    \item{\code{Lung_T1.1034270287.BEAD_STDERR}}{a numeric vector}
    \item{\code{Lung_T10.1034271154.AVG_Beta}}{a numeric vector}
    \item{\code{Lung_T10.1034271154.Signal.CY3}}{a numeric vector}
    \item{\code{Lung_T10.1034271154.Signal.CY5}}{a numeric vector}
    \item{\code{Lung_T10.1034271154.Avg_NBEADS}}{a numeric vector}
    \item{\code{Lung_T10.1034271154.BEAD_STDERR}}{a numeric vector}
    \item{\code{Lung_T11.1034271167.AVG_Beta}}{a numeric vector}
    \item{\code{Lung_T11.1034271167.Signal.CY3}}{a numeric vector}
    \item{\code{Lung_T11.1034271167.Signal.CY5}}{a numeric vector}
    \item{\code{Lung_T11.1034271167.Avg_NBEADS}}{a numeric vector}
    \item{\code{Lung_T11.1034271167.BEAD_STDERR}}{a numeric vector}
    \item{\code{Lung_T12.1034271176.AVG_Beta}}{a numeric vector}
    \item{\code{Lung_T12.1034271176.Signal.CY3}}{a numeric vector}
    \item{\code{Lung_T12.1034271176.Signal.CY5}}{a numeric vector}
    \item{\code{Lung_T12.1034271176.Avg_NBEADS}}{a numeric vector}
    \item{\code{Lung_T12.1034271176.BEAD_STDERR}}{a numeric vector}
    \item{\code{Lung_T13.1034271192.AVG_Beta}}{a numeric vector}
    \item{\code{Lung_T13.1034271192.Signal.CY3}}{a numeric vector}
    \item{\code{Lung_T13.1034271192.Signal.CY5}}{a numeric vector}
    \item{\code{Lung_T13.1034271192.Avg_NBEADS}}{a numeric vector}
    \item{\code{Lung_T13.1034271192.BEAD_STDERR}}{a numeric vector}
    \item{\code{Lung_T14.1034271199.AVG_Beta}}{a numeric vector}
    \item{\code{Lung_T14.1034271199.Signal.CY3}}{a numeric vector}
    \item{\code{Lung_T14.1034271199.Signal.CY5}}{a numeric vector}
    \item{\code{Lung_T14.1034271199.Avg_NBEADS}}{a numeric vector}
    \item{\code{Lung_T14.1034271199.BEAD_STDERR}}{a numeric vector}
    \item{\code{Lung_N15.1034271218.AVG_Beta}}{a numeric vector}
    \item{\code{Lung_N15.1034271218.Signal.CY3}}{a numeric vector}
    \item{\code{Lung_N15.1034271218.Signal.CY5}}{a numeric vector}
    \item{\code{Lung_N15.1034271218.Avg_NBEADS}}{a numeric vector}
    \item{\code{Lung_N15.1034271218.BEAD_STDERR}}{a numeric vector}
    \item{\code{Lung_N16.1034271221.AVG_Beta}}{a numeric vector}
    \item{\code{Lung_N16.1034271221.Signal.CY3}}{a numeric vector}
    \item{\code{Lung_N16.1034271221.Signal.CY5}}{a numeric vector}
    \item{\code{Lung_N16.1034271221.Avg_NBEADS}}{a numeric vector}
    \item{\code{Lung_N16.1034271221.BEAD_STDERR}}{a numeric vector}
    \item{\code{Lung_N17.1034269672.AVG_Beta}}{a numeric vector}
    \item{\code{Lung_N17.1034269672.Signal.CY3}}{a numeric vector}
    \item{\code{Lung_N17.1034269672.Signal.CY5}}{a numeric vector}
    \item{\code{Lung_N17.1034269672.Avg_NBEADS}}{a numeric vector}
    \item{\code{Lung_N17.1034269672.BEAD_STDERR}}{a numeric vector}
    \item{\code{Lung_N18.1034247344.AVG_Beta}}{a numeric vector}
    \item{\code{Lung_N18.1034247344.Signal.CY3}}{a numeric vector}
    \item{\code{Lung_N18.1034247344.Signal.CY5}}{a numeric vector}
    \item{\code{Lung_N18.1034247344.Avg_NBEADS}}{a numeric vector}
    \item{\code{Lung_N18.1034247344.BEAD_STDERR}}{a numeric vector}
    \item{\code{Lung_T19.1034247354.AVG_Beta}}{a numeric vector}
    \item{\code{Lung_T19.1034247354.Signal.CY3}}{a numeric vector}
    \item{\code{Lung_T19.1034247354.Signal.CY5}}{a numeric vector}
    \item{\code{Lung_T19.1034247354.Avg_NBEADS}}{a numeric vector}
    \item{\code{Lung_T19.1034247354.BEAD_STDERR}}{a numeric vector}
    \item{\code{Lung_T2.1034270306.AVG_Beta}}{a numeric vector}
    \item{\code{Lung_T2.1034270306.Signal.CY3}}{a numeric vector}
    \item{\code{Lung_T2.1034270306.Signal.CY5}}{a numeric vector}
    \item{\code{Lung_T2.1034270306.Avg_NBEADS}}{a numeric vector}
    \item{\code{Lung_T2.1034270306.BEAD_STDERR}}{a numeric vector}
    \item{\code{Lung_T20.1034247358.AVG_Beta}}{a numeric vector}
    \item{\code{Lung_T20.1034247358.Signal.CY3}}{a numeric vector}
    \item{\code{Lung_T20.1034247358.Signal.CY5}}{a numeric vector}
    \item{\code{Lung_T20.1034247358.Avg_NBEADS}}{a numeric vector}
    \item{\code{Lung_T20.1034247358.BEAD_STDERR}}{a numeric vector}
    \item{\code{Lung_T21.1034247376.AVG_Beta}}{a numeric vector}
    \item{\code{Lung_T21.1034247376.Signal.CY3}}{a numeric vector}
    \item{\code{Lung_T21.1034247376.Signal.CY5}}{a numeric vector}
    \item{\code{Lung_T21.1034247376.Avg_NBEADS}}{a numeric vector}
    \item{\code{Lung_T21.1034247376.BEAD_STDERR}}{a numeric vector}
    \item{\code{Lung_T22.1034247383.AVG_Beta}}{a numeric vector}
    \item{\code{Lung_T22.1034247383.Signal.CY3}}{a numeric vector}
    \item{\code{Lung_T22.1034247383.Signal.CY5}}{a numeric vector}
    \item{\code{Lung_T22.1034247383.Avg_NBEADS}}{a numeric vector}
    \item{\code{Lung_T22.1034247383.BEAD_STDERR}}{a numeric vector}
    \item{\code{Lung_T23.1034247399.AVG_Beta}}{a numeric vector}
    \item{\code{Lung_T23.1034247399.Signal.CY3}}{a numeric vector}
    \item{\code{Lung_T23.1034247399.Signal.CY5}}{a numeric vector}
    \item{\code{Lung_T23.1034247399.Avg_NBEADS}}{a numeric vector}
    \item{\code{Lung_T23.1034247399.BEAD_STDERR}}{a numeric vector}
    \item{\code{Lung_T24.1034247412.AVG_Beta}}{a numeric vector}
    \item{\code{Lung_T24.1034247412.Signal.CY3}}{a numeric vector}
    \item{\code{Lung_T24.1034247412.Signal.CY5}}{a numeric vector}
    \item{\code{Lung_T24.1034247412.Avg_NBEADS}}{a numeric vector}
    \item{\code{Lung_T24.1034247412.BEAD_STDERR}}{a numeric vector}
    \item{\code{Lung_T25.1034247404.AVG_Beta}}{a numeric vector}
    \item{\code{Lung_T25.1034247404.Signal.CY3}}{a numeric vector}
    \item{\code{Lung_T25.1034247404.Signal.CY5}}{a numeric vector}
    \item{\code{Lung_T25.1034247404.Avg_NBEADS}}{a numeric vector}
    \item{\code{Lung_T25.1034247404.BEAD_STDERR}}{a numeric vector}
    \item{\code{Lung_T3.1034270309.AVG_Beta}}{a numeric vector}
    \item{\code{Lung_T3.1034270309.Signal.CY3}}{a numeric vector}
    \item{\code{Lung_T3.1034270309.Signal.CY5}}{a numeric vector}
    \item{\code{Lung_T3.1034270309.Avg_NBEADS}}{a numeric vector}
    \item{\code{Lung_T3.1034270309.BEAD_STDERR}}{a numeric vector}
    \item{\code{Lung_T4.1034270331.AVG_Beta}}{a numeric vector}
    \item{\code{Lung_T4.1034270331.Signal.CY3}}{a numeric vector}
    \item{\code{Lung_T4.1034270331.Signal.CY5}}{a numeric vector}
    \item{\code{Lung_T4.1034270331.Avg_NBEADS}}{a numeric vector}
    \item{\code{Lung_T4.1034270331.BEAD_STDERR}}{a numeric vector}
    \item{\code{Lung_T5.1034270321.AVG_Beta}}{a numeric vector}
    \item{\code{Lung_T5.1034270321.Signal.CY3}}{a numeric vector}
    \item{\code{Lung_T5.1034270321.Signal.CY5}}{a numeric vector}
    \item{\code{Lung_T5.1034270321.Avg_NBEADS}}{a numeric vector}
    \item{\code{Lung_T5.1034270321.BEAD_STDERR}}{a numeric vector}
    \item{\code{Lung_T7.1034270348.AVG_Beta}}{a numeric vector}
    \item{\code{Lung_T7.1034270348.Signal.CY3}}{a numeric vector}
    \item{\code{Lung_T7.1034270348.Signal.CY5}}{a numeric vector}
    \item{\code{Lung_T7.1034270348.Avg_NBEADS}}{a numeric vector}
    \item{\code{Lung_T7.1034270348.BEAD_STDERR}}{a numeric vector}
    \item{\code{Lung_T8.1034270362.AVG_Beta}}{a numeric vector}
    \item{\code{Lung_T8.1034270362.Signal.CY3}}{a numeric vector}
    \item{\code{Lung_T8.1034270362.Signal.CY5}}{a numeric vector}
    \item{\code{Lung_T8.1034270362.Avg_NBEADS}}{a numeric vector}
    \item{\code{Lung_T8.1034270362.BEAD_STDERR}}{a numeric vector}
    \item{\code{Lung_T9.1034269644.AVG_Beta}}{a numeric vector}
    \item{\code{Lung_T9.1034269644.Signal.CY3}}{a numeric vector}
    \item{\code{Lung_T9.1034269644.Signal.CY5}}{a numeric vector}
    \item{\code{Lung_T9.1034269644.Avg_NBEADS}}{a numeric vector}
    \item{\code{Lung_T9.1034269644.BEAD_STDERR}}{a numeric vector}
    \item{\code{Wilms_N1.1034270294.AVG_Beta}}{a numeric vector}
    \item{\code{Wilms_N1.1034270294.Signal.CY3}}{a numeric vector}
    \item{\code{Wilms_N1.1034270294.Signal.CY5}}{a numeric vector}
    \item{\code{Wilms_N1.1034270294.Avg_NBEADS}}{a numeric vector}
    \item{\code{Wilms_N1.1034270294.BEAD_STDERR}}{a numeric vector}
    \item{\code{Wilms_N10.1034271170.AVG_Beta}}{a numeric vector}
    \item{\code{Wilms_N10.1034271170.Signal.CY3}}{a numeric vector}
    \item{\code{Wilms_N10.1034271170.Signal.CY5}}{a numeric vector}
    \item{\code{Wilms_N10.1034271170.Avg_NBEADS}}{a numeric vector}
    \item{\code{Wilms_N10.1034271170.BEAD_STDERR}}{a numeric vector}
    \item{\code{Wilms_N11.1034271172.AVG_Beta}}{a numeric vector}
    \item{\code{Wilms_N11.1034271172.Signal.CY3}}{a numeric vector}
    \item{\code{Wilms_N11.1034271172.Signal.CY5}}{a numeric vector}
    \item{\code{Wilms_N11.1034271172.Avg_NBEADS}}{a numeric vector}
    \item{\code{Wilms_N11.1034271172.BEAD_STDERR}}{a numeric vector}
    \item{\code{Wilms_N12.1034271195.AVG_Beta}}{a numeric vector}
    \item{\code{Wilms_N12.1034271195.Signal.CY3}}{a numeric vector}
    \item{\code{Wilms_N12.1034271195.Signal.CY5}}{a numeric vector}
    \item{\code{Wilms_N12.1034271195.Avg_NBEADS}}{a numeric vector}
    \item{\code{Wilms_N12.1034271195.BEAD_STDERR}}{a numeric vector}
    \item{\code{Wilms_N13.1034271196.AVG_Beta}}{a numeric vector}
    \item{\code{Wilms_N13.1034271196.Signal.CY3}}{a numeric vector}
    \item{\code{Wilms_N13.1034271196.Signal.CY5}}{a numeric vector}
    \item{\code{Wilms_N13.1034271196.Avg_NBEADS}}{a numeric vector}
    \item{\code{Wilms_N13.1034271196.BEAD_STDERR}}{a numeric vector}
    \item{\code{Wilms_N14.1034271206.AVG_Beta}}{a numeric vector}
    \item{\code{Wilms_N14.1034271206.Signal.CY3}}{a numeric vector}
    \item{\code{Wilms_N14.1034271206.Signal.CY5}}{a numeric vector}
    \item{\code{Wilms_N14.1034271206.Avg_NBEADS}}{a numeric vector}
    \item{\code{Wilms_N14.1034271206.BEAD_STDERR}}{a numeric vector}
    \item{\code{Wilms_N15.1034271210.AVG_Beta}}{a numeric vector}
    \item{\code{Wilms_N15.1034271210.Signal.CY3}}{a numeric vector}
    \item{\code{Wilms_N15.1034271210.Signal.CY5}}{a numeric vector}
    \item{\code{Wilms_N15.1034271210.Avg_NBEADS}}{a numeric vector}
    \item{\code{Wilms_N15.1034271210.BEAD_STDERR}}{a numeric vector}
    \item{\code{Wilms_N16.1034269691.AVG_Beta}}{a numeric vector}
    \item{\code{Wilms_N16.1034269691.Signal.CY3}}{a numeric vector}
    \item{\code{Wilms_N16.1034269691.Signal.CY5}}{a numeric vector}
    \item{\code{Wilms_N16.1034269691.Avg_NBEADS}}{a numeric vector}
    \item{\code{Wilms_N16.1034269691.BEAD_STDERR}}{a numeric vector}
    \item{\code{Wilms_N17.1034271237.AVG_Beta}}{a numeric vector}
    \item{\code{Wilms_N17.1034271237.Signal.CY3}}{a numeric vector}
    \item{\code{Wilms_N17.1034271237.Signal.CY5}}{a numeric vector}
    \item{\code{Wilms_N17.1034271237.Avg_NBEADS}}{a numeric vector}
    \item{\code{Wilms_N17.1034271237.BEAD_STDERR}}{a numeric vector}
    \item{\code{Wilms_N18.1034247351.AVG_Beta}}{a numeric vector}
    \item{\code{Wilms_N18.1034247351.Signal.CY3}}{a numeric vector}
    \item{\code{Wilms_N18.1034247351.Signal.CY5}}{a numeric vector}
    \item{\code{Wilms_N18.1034247351.Avg_NBEADS}}{a numeric vector}
    \item{\code{Wilms_N18.1034247351.BEAD_STDERR}}{a numeric vector}
    \item{\code{Wilms_N19.1034247361.AVG_Beta}}{a numeric vector}
    \item{\code{Wilms_N19.1034247361.Signal.CY3}}{a numeric vector}
    \item{\code{Wilms_N19.1034247361.Signal.CY5}}{a numeric vector}
    \item{\code{Wilms_N19.1034247361.Avg_NBEADS}}{a numeric vector}
    \item{\code{Wilms_N19.1034247361.BEAD_STDERR}}{a numeric vector}
    \item{\code{Wilms_N2.1034270299.AVG_Beta}}{a numeric vector}
    \item{\code{Wilms_N2.1034270299.Signal.CY3}}{a numeric vector}
    \item{\code{Wilms_N2.1034270299.Signal.CY5}}{a numeric vector}
    \item{\code{Wilms_N2.1034270299.Avg_NBEADS}}{a numeric vector}
    \item{\code{Wilms_N2.1034270299.BEAD_STDERR}}{a numeric vector}
    \item{\code{Wilms_N20.1034247373.AVG_Beta}}{a numeric vector}
    \item{\code{Wilms_N20.1034247373.Signal.CY3}}{a numeric vector}
    \item{\code{Wilms_N20.1034247373.Signal.CY5}}{a numeric vector}
    \item{\code{Wilms_N20.1034247373.Avg_NBEADS}}{a numeric vector}
    \item{\code{Wilms_N20.1034247373.BEAD_STDERR}}{a numeric vector}
    \item{\code{Wilms_N21.1034247387.AVG_Beta}}{a numeric vector}
    \item{\code{Wilms_N21.1034247387.Signal.CY3}}{a numeric vector}
    \item{\code{Wilms_N21.1034247387.Signal.CY5}}{a numeric vector}
    \item{\code{Wilms_N21.1034247387.Avg_NBEADS}}{a numeric vector}
    \item{\code{Wilms_N21.1034247387.BEAD_STDERR}}{a numeric vector}
    \item{\code{Wilms_N22.1034247396.AVG_Beta}}{a numeric vector}
    \item{\code{Wilms_N22.1034247396.Signal.CY3}}{a numeric vector}
    \item{\code{Wilms_N22.1034247396.Signal.CY5}}{a numeric vector}
    \item{\code{Wilms_N22.1034247396.Avg_NBEADS}}{a numeric vector}
    \item{\code{Wilms_N22.1034247396.BEAD_STDERR}}{a numeric vector}
    \item{\code{Wilms_N23.1034247413.AVG_Beta}}{a numeric vector}
    \item{\code{Wilms_N23.1034247413.Signal.CY3}}{a numeric vector}
    \item{\code{Wilms_N23.1034247413.Signal.CY5}}{a numeric vector}
    \item{\code{Wilms_N23.1034247413.Avg_NBEADS}}{a numeric vector}
    \item{\code{Wilms_N23.1034247413.BEAD_STDERR}}{a numeric vector}
    \item{\code{Wilms_N24.1034247405.AVG_Beta}}{a numeric vector}
    \item{\code{Wilms_N24.1034247405.Signal.CY3}}{a numeric vector}
    \item{\code{Wilms_N24.1034247405.Signal.CY5}}{a numeric vector}
    \item{\code{Wilms_N24.1034247405.Avg_NBEADS}}{a numeric vector}
    \item{\code{Wilms_N24.1034247405.BEAD_STDERR}}{a numeric vector}
    \item{\code{Wilms_N25.1034247424.AVG_Beta}}{a numeric vector}
    \item{\code{Wilms_N25.1034247424.Signal.CY3}}{a numeric vector}
    \item{\code{Wilms_N25.1034247424.Signal.CY5}}{a numeric vector}
    \item{\code{Wilms_N25.1034247424.Avg_NBEADS}}{a numeric vector}
    \item{\code{Wilms_N25.1034247424.BEAD_STDERR}}{a numeric vector}
    \item{\code{Wilms_N26.1034247436.AVG_Beta}}{a numeric vector}
    \item{\code{Wilms_N26.1034247436.Signal.CY3}}{a numeric vector}
    \item{\code{Wilms_N26.1034247436.Signal.CY5}}{a numeric vector}
    \item{\code{Wilms_N26.1034247436.Avg_NBEADS}}{a numeric vector}
    \item{\code{Wilms_N26.1034247436.BEAD_STDERR}}{a numeric vector}
    \item{\code{Wilms_N3.1034270316.AVG_Beta}}{a numeric vector}
    \item{\code{Wilms_N3.1034270316.Signal.CY3}}{a numeric vector}
    \item{\code{Wilms_N3.1034270316.Signal.CY5}}{a numeric vector}
    \item{\code{Wilms_N3.1034270316.Avg_NBEADS}}{a numeric vector}
    \item{\code{Wilms_N3.1034270316.BEAD_STDERR}}{a numeric vector}
    \item{\code{Wilms_N4.1034270324.AVG_Beta}}{a numeric vector}
    \item{\code{Wilms_N4.1034270324.Signal.CY3}}{a numeric vector}
    \item{\code{Wilms_N4.1034270324.Signal.CY5}}{a numeric vector}
    \item{\code{Wilms_N4.1034270324.Avg_NBEADS}}{a numeric vector}
    \item{\code{Wilms_N4.1034270324.BEAD_STDERR}}{a numeric vector}
    \item{\code{Wilms_N5.1034270338.AVG_Beta}}{a numeric vector}
    \item{\code{Wilms_N5.1034270338.Signal.CY3}}{a numeric vector}
    \item{\code{Wilms_N5.1034270338.Signal.CY5}}{a numeric vector}
    \item{\code{Wilms_N5.1034270338.Avg_NBEADS}}{a numeric vector}
    \item{\code{Wilms_N5.1034270338.BEAD_STDERR}}{a numeric vector}
    \item{\code{Wilms_N6.1034270351.AVG_Beta}}{a numeric vector}
    \item{\code{Wilms_N6.1034270351.Signal.CY3}}{a numeric vector}
    \item{\code{Wilms_N6.1034270351.Signal.CY5}}{a numeric vector}
    \item{\code{Wilms_N6.1034270351.Avg_NBEADS}}{a numeric vector}
    \item{\code{Wilms_N6.1034270351.BEAD_STDERR}}{a numeric vector}
    \item{\code{Wilms_N8.1034270376.AVG_Beta}}{a numeric vector}
    \item{\code{Wilms_N8.1034270376.Signal.CY3}}{a numeric vector}
    \item{\code{Wilms_N8.1034270376.Signal.CY5}}{a numeric vector}
    \item{\code{Wilms_N8.1034270376.Avg_NBEADS}}{a numeric vector}
    \item{\code{Wilms_N8.1034270376.BEAD_STDERR}}{a numeric vector}
    \item{\code{Wilms_N9.1034271151.AVG_Beta}}{a numeric vector}
    \item{\code{Wilms_N9.1034271151.Signal.CY3}}{a numeric vector}
    \item{\code{Wilms_N9.1034271151.Signal.CY5}}{a numeric vector}
    \item{\code{Wilms_N9.1034271151.Avg_NBEADS}}{a numeric vector}
    \item{\code{Wilms_N9.1034271151.BEAD_STDERR}}{a numeric vector}
    \item{\code{Wilms_T1.1034270293.AVG_Beta}}{a numeric vector}
    \item{\code{Wilms_T1.1034270293.Signal.CY3}}{a numeric vector}
    \item{\code{Wilms_T1.1034270293.Signal.CY5}}{a numeric vector}
    \item{\code{Wilms_T1.1034270293.Avg_NBEADS}}{a numeric vector}
    \item{\code{Wilms_T1.1034270293.BEAD_STDERR}}{a numeric vector}
    \item{\code{Wilms_T10.1034269694.AVG_Beta}}{a numeric vector}
    \item{\code{Wilms_T10.1034269694.Signal.CY3}}{a numeric vector}
    \item{\code{Wilms_T10.1034269694.Signal.CY5}}{a numeric vector}
    \item{\code{Wilms_T10.1034269694.Avg_NBEADS}}{a numeric vector}
    \item{\code{Wilms_T10.1034269694.BEAD_STDERR}}{a numeric vector}
    \item{\code{Wilms_T11.1034271173.AVG_Beta}}{a numeric vector}
    \item{\code{Wilms_T11.1034271173.Signal.CY3}}{a numeric vector}
    \item{\code{Wilms_T11.1034271173.Signal.CY5}}{a numeric vector}
    \item{\code{Wilms_T11.1034271173.Avg_NBEADS}}{a numeric vector}
    \item{\code{Wilms_T11.1034271173.BEAD_STDERR}}{a numeric vector}
    \item{\code{Wilms_T12.1034271183.AVG_Beta}}{a numeric vector}
    \item{\code{Wilms_T12.1034271183.Signal.CY3}}{a numeric vector}
    \item{\code{Wilms_T12.1034271183.Signal.CY5}}{a numeric vector}
    \item{\code{Wilms_T12.1034271183.Avg_NBEADS}}{a numeric vector}
    \item{\code{Wilms_T12.1034271183.BEAD_STDERR}}{a numeric vector}
    \item{\code{Wilms_T13.1034271185.AVG_Beta}}{a numeric vector}
    \item{\code{Wilms_T13.1034271185.Signal.CY3}}{a numeric vector}
    \item{\code{Wilms_T13.1034271185.Signal.CY5}}{a numeric vector}
    \item{\code{Wilms_T13.1034271185.Avg_NBEADS}}{a numeric vector}
    \item{\code{Wilms_T13.1034271185.BEAD_STDERR}}{a numeric vector}
    \item{\code{Wilms_T14.1034271205.AVG_Beta}}{a numeric vector}
    \item{\code{Wilms_T14.1034271205.Signal.CY3}}{a numeric vector}
    \item{\code{Wilms_T14.1034271205.Signal.CY5}}{a numeric vector}
    \item{\code{Wilms_T14.1034271205.Avg_NBEADS}}{a numeric vector}
    \item{\code{Wilms_T14.1034271205.BEAD_STDERR}}{a numeric vector}
    \item{\code{Wilms_T15.1034271211.AVG_Beta}}{a numeric vector}
    \item{\code{Wilms_T15.1034271211.Signal.CY3}}{a numeric vector}
    \item{\code{Wilms_T15.1034271211.Signal.CY5}}{a numeric vector}
    \item{\code{Wilms_T15.1034271211.Avg_NBEADS}}{a numeric vector}
    \item{\code{Wilms_T15.1034271211.BEAD_STDERR}}{a numeric vector}
    \item{\code{Wilms_T16.1034271226.AVG_Beta}}{a numeric vector}
    \item{\code{Wilms_T16.1034271226.Signal.CY3}}{a numeric vector}
    \item{\code{Wilms_T16.1034271226.Signal.CY5}}{a numeric vector}
    \item{\code{Wilms_T16.1034271226.Avg_NBEADS}}{a numeric vector}
    \item{\code{Wilms_T16.1034271226.BEAD_STDERR}}{a numeric vector}
    \item{\code{Wilms_T17.1034271238.AVG_Beta}}{a numeric vector}
    \item{\code{Wilms_T17.1034271238.Signal.CY3}}{a numeric vector}
    \item{\code{Wilms_T17.1034271238.Signal.CY5}}{a numeric vector}
    \item{\code{Wilms_T17.1034271238.Avg_NBEADS}}{a numeric vector}
    \item{\code{Wilms_T17.1034271238.BEAD_STDERR}}{a numeric vector}
    \item{\code{Wilms_T18.1034247350.AVG_Beta}}{a numeric vector}
    \item{\code{Wilms_T18.1034247350.Signal.CY3}}{a numeric vector}
    \item{\code{Wilms_T18.1034247350.Signal.CY5}}{a numeric vector}
    \item{\code{Wilms_T18.1034247350.Avg_NBEADS}}{a numeric vector}
    \item{\code{Wilms_T18.1034247350.BEAD_STDERR}}{a numeric vector}
    \item{\code{Wilms_T19.1034247362.AVG_Beta}}{a numeric vector}
    \item{\code{Wilms_T19.1034247362.Signal.CY3}}{a numeric vector}
    \item{\code{Wilms_T19.1034247362.Signal.CY5}}{a numeric vector}
    \item{\code{Wilms_T19.1034247362.Avg_NBEADS}}{a numeric vector}
    \item{\code{Wilms_T19.1034247362.BEAD_STDERR}}{a numeric vector}
    \item{\code{Wilms_T2.1034270300.AVG_Beta}}{a numeric vector}
    \item{\code{Wilms_T2.1034270300.Signal.CY3}}{a numeric vector}
    \item{\code{Wilms_T2.1034270300.Signal.CY5}}{a numeric vector}
    \item{\code{Wilms_T2.1034270300.Avg_NBEADS}}{a numeric vector}
    \item{\code{Wilms_T2.1034270300.BEAD_STDERR}}{a numeric vector}
    \item{\code{Wilms_T20.1034247372.AVG_Beta}}{a numeric vector}
    \item{\code{Wilms_T20.1034247372.Signal.CY3}}{a numeric vector}
    \item{\code{Wilms_T20.1034247372.Signal.CY5}}{a numeric vector}
    \item{\code{Wilms_T20.1034247372.Avg_NBEADS}}{a numeric vector}
    \item{\code{Wilms_T20.1034247372.BEAD_STDERR}}{a numeric vector}
    \item{\code{Wilms_T21.1034247388.AVG_Beta}}{a numeric vector}
    \item{\code{Wilms_T21.1034247388.Signal.CY3}}{a numeric vector}
    \item{\code{Wilms_T21.1034247388.Signal.CY5}}{a numeric vector}
    \item{\code{Wilms_T21.1034247388.Avg_NBEADS}}{a numeric vector}
    \item{\code{Wilms_T21.1034247388.BEAD_STDERR}}{a numeric vector}
    \item{\code{Wilms_T22.1034247395.AVG_Beta}}{a numeric vector}
    \item{\code{Wilms_T22.1034247395.Signal.CY3}}{a numeric vector}
    \item{\code{Wilms_T22.1034247395.Signal.CY5}}{a numeric vector}
    \item{\code{Wilms_T22.1034247395.Avg_NBEADS}}{a numeric vector}
    \item{\code{Wilms_T22.1034247395.BEAD_STDERR}}{a numeric vector}
    \item{\code{Wilms_T23.1034247414.AVG_Beta}}{a numeric vector}
    \item{\code{Wilms_T23.1034247414.Signal.CY3}}{a numeric vector}
    \item{\code{Wilms_T23.1034247414.Signal.CY5}}{a numeric vector}
    \item{\code{Wilms_T23.1034247414.Avg_NBEADS}}{a numeric vector}
    \item{\code{Wilms_T23.1034247414.BEAD_STDERR}}{a numeric vector}
    \item{\code{Wilms_T24.1034247406.AVG_Beta}}{a numeric vector}
    \item{\code{Wilms_T24.1034247406.Signal.CY3}}{a numeric vector}
    \item{\code{Wilms_T24.1034247406.Signal.CY5}}{a numeric vector}
    \item{\code{Wilms_T24.1034247406.Avg_NBEADS}}{a numeric vector}
    \item{\code{Wilms_T24.1034247406.BEAD_STDERR}}{a numeric vector}
    \item{\code{Wilms_T25.1034247422.AVG_Beta}}{a numeric vector}
    \item{\code{Wilms_T25.1034247422.Signal.CY3}}{a numeric vector}
    \item{\code{Wilms_T25.1034247422.Signal.CY5}}{a numeric vector}
    \item{\code{Wilms_T25.1034247422.Avg_NBEADS}}{a numeric vector}
    \item{\code{Wilms_T25.1034247422.BEAD_STDERR}}{a numeric vector}
    \item{\code{Wilms_T26.1034247437.AVG_Beta}}{a numeric vector}
    \item{\code{Wilms_T26.1034247437.Signal.CY3}}{a numeric vector}
    \item{\code{Wilms_T26.1034247437.Signal.CY5}}{a numeric vector}
    \item{\code{Wilms_T26.1034247437.Avg_NBEADS}}{a numeric vector}
    \item{\code{Wilms_T26.1034247437.BEAD_STDERR}}{a numeric vector}
    \item{\code{Wilms_T3.1034270315.AVG_Beta}}{a numeric vector}
    \item{\code{Wilms_T3.1034270315.Signal.CY3}}{a numeric vector}
    \item{\code{Wilms_T3.1034270315.Signal.CY5}}{a numeric vector}
    \item{\code{Wilms_T3.1034270315.Avg_NBEADS}}{a numeric vector}
    \item{\code{Wilms_T3.1034270315.BEAD_STDERR}}{a numeric vector}
    \item{\code{Wilms_T4.1034270325.AVG_Beta}}{a numeric vector}
    \item{\code{Wilms_T4.1034270325.Signal.CY3}}{a numeric vector}
    \item{\code{Wilms_T4.1034270325.Signal.CY5}}{a numeric vector}
    \item{\code{Wilms_T4.1034270325.Avg_NBEADS}}{a numeric vector}
    \item{\code{Wilms_T4.1034270325.BEAD_STDERR}}{a numeric vector}
    \item{\code{Wilms_T5.1034270337.AVG_Beta}}{a numeric vector}
    \item{\code{Wilms_T5.1034270337.Signal.CY3}}{a numeric vector}
    \item{\code{Wilms_T5.1034270337.Signal.CY5}}{a numeric vector}
    \item{\code{Wilms_T5.1034270337.Avg_NBEADS}}{a numeric vector}
    \item{\code{Wilms_T5.1034270337.BEAD_STDERR}}{a numeric vector}
    \item{\code{Wilms_T6.1034270352.AVG_Beta}}{a numeric vector}
    \item{\code{Wilms_T6.1034270352.Signal.CY3}}{a numeric vector}
    \item{\code{Wilms_T6.1034270352.Signal.CY5}}{a numeric vector}
    \item{\code{Wilms_T6.1034270352.Avg_NBEADS}}{a numeric vector}
    \item{\code{Wilms_T6.1034270352.BEAD_STDERR}}{a numeric vector}
    \item{\code{Wilms_T8.1034270377.AVG_Beta}}{a numeric vector}
    \item{\code{Wilms_T8.1034270377.Signal.CY3}}{a numeric vector}
    \item{\code{Wilms_T8.1034270377.Signal.CY5}}{a numeric vector}
    \item{\code{Wilms_T8.1034270377.Avg_NBEADS}}{a numeric vector}
    \item{\code{Wilms_T8.1034270377.BEAD_STDERR}}{a numeric vector}
    \item{\code{Wilms_T9.1034269698.AVG_Beta}}{a numeric vector}
    \item{\code{Wilms_T9.1034269698.Signal.CY3}}{a numeric vector}
    \item{\code{Wilms_T9.1034269698.Signal.CY5}}{a numeric vector}
    \item{\code{Wilms_T9.1034269698.Avg_NBEADS}}{a numeric vector}
    \item{\code{Wilms_T9.1034269698.BEAD_STDERR}}{a numeric vector}
    \item{\code{ThyPTC_T5.0098609768.AVG_Beta}}{a numeric vector}
    \item{\code{ThyPTC_T5.0098609768.Signal.CY3}}{a numeric vector}
    \item{\code{ThyPTC_T5.0098609768.Signal.CY5}}{a numeric vector}
    \item{\code{ThyPTC_T5.0098609768.Avg_NBEADS}}{a numeric vector}
    \item{\code{ThyPTC_T5.0098609768.BEAD_STDERR}}{a numeric vector}
    \item{\code{ThyPTC_N5.0098609752.AVG_Beta}}{a numeric vector}
    \item{\code{ThyPTC_N5.0098609752.Signal.CY3}}{a numeric vector}
    \item{\code{ThyPTC_N5.0098609752.Signal.CY5}}{a numeric vector}
    \item{\code{ThyPTC_N5.0098609752.Avg_NBEADS}}{a numeric vector}
    \item{\code{ThyPTC_N5.0098609752.BEAD_STDERR}}{a numeric vector}
    \item{\code{ThyFA_N6.0098612947.AVG_Beta}}{a numeric vector}
    \item{\code{ThyFA_N6.0098612947.Signal.CY3}}{a numeric vector}
    \item{\code{ThyFA_N6.0098612947.Signal.CY5}}{a numeric vector}
    \item{\code{ThyFA_N6.0098612947.Avg_NBEADS}}{a numeric vector}
    \item{\code{ThyFA_N6.0098612947.BEAD_STDERR}}{a numeric vector}
    \item{\code{ThyFC_T6.0098612948.AVG_Beta}}{a numeric vector}
    \item{\code{ThyFC_T6.0098612948.Signal.CY3}}{a numeric vector}
    \item{\code{ThyFC_T6.0098612948.Signal.CY5}}{a numeric vector}
    \item{\code{ThyFC_T6.0098612948.Avg_NBEADS}}{a numeric vector}
    \item{\code{ThyFC_T6.0098612948.BEAD_STDERR}}{a numeric vector}
    \item{\code{ThyFVPTC_T6.0098612949.AVG_Beta}}{a numeric vector}
    \item{\code{ThyFVPTC_T6.0098612949.Signal.CY3}}{a numeric vector}
    \item{\code{ThyFVPTC_T6.0098612949.Signal.CY5}}{a numeric vector}
    \item{\code{ThyFVPTC_T6.0098612949.Avg_NBEADS}}{a numeric vector}
    \item{\code{ThyFVPTC_T6.0098612949.BEAD_STDERR}}{a numeric vector}
    \item{\code{ThyFVPTC_N6.0098612950.AVG_Beta}}{a numeric vector}
    \item{\code{ThyFVPTC_N6.0098612950.Signal.CY3}}{a numeric vector}
    \item{\code{ThyFVPTC_N6.0098612950.Signal.CY5}}{a numeric vector}
    \item{\code{ThyFVPTC_N6.0098612950.Avg_NBEADS}}{a numeric vector}
    \item{\code{ThyFVPTC_N6.0098612950.BEAD_STDERR}}{a numeric vector}
    \item{\code{Adenoma_T7.0098612937.AVG_Beta}}{a numeric vector}
    \item{\code{Adenoma_T7.0098612937.Signal.CY3}}{a numeric vector}
    \item{\code{Adenoma_T7.0098612937.Signal.CY5}}{a numeric vector}
    \item{\code{Adenoma_T7.0098612937.Avg_NBEADS}}{a numeric vector}
    \item{\code{Adenoma_T7.0098612937.BEAD_STDERR}}{a numeric vector}
    \item{\code{ThyPTC_N7.0098612923.AVG_Beta}}{a numeric vector}
    \item{\code{ThyPTC_N7.0098612923.Signal.CY3}}{a numeric vector}
    \item{\code{ThyPTC_N7.0098612923.Signal.CY5}}{a numeric vector}
    \item{\code{ThyPTC_N7.0098612923.Avg_NBEADS}}{a numeric vector}
    \item{\code{ThyPTC_N7.0098612923.BEAD_STDERR}}{a numeric vector}
    \item{\code{Adenoma_T8.0098612925.AVG_Beta}}{a numeric vector}
    \item{\code{Adenoma_T8.0098612925.Signal.CY3}}{a numeric vector}
    \item{\code{Adenoma_T8.0098612925.Signal.CY5}}{a numeric vector}
    \item{\code{Adenoma_T8.0098612925.Avg_NBEADS}}{a numeric vector}
    \item{\code{Adenoma_T8.0098612925.BEAD_STDERR}}{a numeric vector}
    \item{\code{Adenoma_T9.0098612914.AVG_Beta}}{a numeric vector}
    \item{\code{Adenoma_T9.0098612914.Signal.CY3}}{a numeric vector}
    \item{\code{Adenoma_T9.0098612914.Signal.CY5}}{a numeric vector}
    \item{\code{Adenoma_T9.0098612914.Avg_NBEADS}}{a numeric vector}
    \item{\code{Adenoma_T9.0098612914.BEAD_STDERR}}{a numeric vector}
    \item{\code{ThyFA_N9.0098612911.AVG_Beta}}{a numeric vector}
    \item{\code{ThyFA_N9.0098612911.Signal.CY3}}{a numeric vector}
    \item{\code{ThyFA_N9.0098612911.Signal.CY5}}{a numeric vector}
    \item{\code{ThyFA_N9.0098612911.Avg_NBEADS}}{a numeric vector}
    \item{\code{ThyFA_N9.0098612911.BEAD_STDERR}}{a numeric vector}
    \item{\code{ThyFA_N10.0098612901.AVG_Beta}}{a numeric vector}
    \item{\code{ThyFA_N10.0098612901.Signal.CY3}}{a numeric vector}
    \item{\code{ThyFA_N10.0098612901.Signal.CY5}}{a numeric vector}
    \item{\code{ThyFA_N10.0098612901.Avg_NBEADS}}{a numeric vector}
    \item{\code{ThyFA_N10.0098612901.BEAD_STDERR}}{a numeric vector}
    \item{\code{Adenoma_T14.0098612890.AVG_Beta}}{a numeric vector}
    \item{\code{Adenoma_T14.0098612890.Signal.CY3}}{a numeric vector}
    \item{\code{Adenoma_T14.0098612890.Signal.CY5}}{a numeric vector}
    \item{\code{Adenoma_T14.0098612890.Avg_NBEADS}}{a numeric vector}
    \item{\code{Adenoma_T14.0098612890.BEAD_STDERR}}{a numeric vector}
    \item{\code{Adenoma_T15.0098612889.AVG_Beta}}{a numeric vector}
    \item{\code{Adenoma_T15.0098612889.Signal.CY3}}{a numeric vector}
    \item{\code{Adenoma_T15.0098612889.Signal.CY5}}{a numeric vector}
    \item{\code{Adenoma_T15.0098612889.Avg_NBEADS}}{a numeric vector}
    \item{\code{Adenoma_T15.0098612889.BEAD_STDERR}}{a numeric vector}
    \item{\code{Adenoma_T16.0098612888.AVG_Beta}}{a numeric vector}
    \item{\code{Adenoma_T16.0098612888.Signal.CY3}}{a numeric vector}
    \item{\code{Adenoma_T16.0098612888.Signal.CY5}}{a numeric vector}
    \item{\code{Adenoma_T16.0098612888.Avg_NBEADS}}{a numeric vector}
    \item{\code{Adenoma_T16.0098612888.BEAD_STDERR}}{a numeric vector}
    \item{\code{Adenoma_T17.0098612887.AVG_Beta}}{a numeric vector}
    \item{\code{Adenoma_T17.0098612887.Signal.CY3}}{a numeric vector}
    \item{\code{Adenoma_T17.0098612887.Signal.CY5}}{a numeric vector}
    \item{\code{Adenoma_T17.0098612887.Avg_NBEADS}}{a numeric vector}
    \item{\code{Adenoma_T17.0098612887.BEAD_STDERR}}{a numeric vector}
    \item{\code{Breast_N30.0098612875.AVG_Beta}}{a numeric vector}
    \item{\code{Breast_N30.0098612875.Signal.CY3}}{a numeric vector}
    \item{\code{Breast_N30.0098612875.Signal.CY5}}{a numeric vector}
    \item{\code{Breast_N30.0098612875.Avg_NBEADS}}{a numeric vector}
    \item{\code{Breast_N30.0098612875.BEAD_STDERR}}{a numeric vector}
    \item{\code{Colon_N31.0098612876.AVG_Beta}}{a numeric vector}
    \item{\code{Colon_N31.0098612876.Signal.CY3}}{a numeric vector}
    \item{\code{Colon_N31.0098612876.Signal.CY5}}{a numeric vector}
    \item{\code{Colon_N31.0098612876.Avg_NBEADS}}{a numeric vector}
    \item{\code{Colon_N31.0098612876.BEAD_STDERR}}{a numeric vector}
    \item{\code{Colon_T31.0098612877.AVG_Beta}}{a numeric vector}
    \item{\code{Colon_T31.0098612877.Signal.CY3}}{a numeric vector}
    \item{\code{Colon_T31.0098612877.Signal.CY5}}{a numeric vector}
    \item{\code{Colon_T31.0098612877.Avg_NBEADS}}{a numeric vector}
    \item{\code{Colon_T31.0098612877.BEAD_STDERR}}{a numeric vector}
    \item{\code{Breast_T31.0098612878.AVG_Beta}}{a numeric vector}
    \item{\code{Breast_T31.0098612878.Signal.CY3}}{a numeric vector}
    \item{\code{Breast_T31.0098612878.Signal.CY5}}{a numeric vector}
    \item{\code{Breast_T31.0098612878.Avg_NBEADS}}{a numeric vector}
    \item{\code{Breast_T31.0098612878.BEAD_STDERR}}{a numeric vector}
    \item{\code{ThyPTC_T6.0098612953.AVG_Beta}}{a numeric vector}
    \item{\code{ThyPTC_T6.0098612953.Signal.CY3}}{a numeric vector}
    \item{\code{ThyPTC_T6.0098612953.Signal.CY5}}{a numeric vector}
    \item{\code{ThyPTC_T6.0098612953.Avg_NBEADS}}{a numeric vector}
    \item{\code{ThyPTC_T6.0098612953.BEAD_STDERR}}{a numeric vector}
    \item{\code{ThyPTC_N6.0098612954.AVG_Beta}}{a numeric vector}
    \item{\code{ThyPTC_N6.0098612954.Signal.CY3}}{a numeric vector}
    \item{\code{ThyPTC_N6.0098612954.Signal.CY5}}{a numeric vector}
    \item{\code{ThyPTC_N6.0098612954.Avg_NBEADS}}{a numeric vector}
    \item{\code{ThyPTC_N6.0098612954.BEAD_STDERR}}{a numeric vector}
    \item{\code{ThyFA_N7.0098612933.AVG_Beta}}{a numeric vector}
    \item{\code{ThyFA_N7.0098612933.Signal.CY3}}{a numeric vector}
    \item{\code{ThyFA_N7.0098612933.Signal.CY5}}{a numeric vector}
    \item{\code{ThyFA_N7.0098612933.Avg_NBEADS}}{a numeric vector}
    \item{\code{ThyFA_N7.0098612933.BEAD_STDERR}}{a numeric vector}
    \item{\code{ThyFC_T7.0098612932.AVG_Beta}}{a numeric vector}
    \item{\code{ThyFC_T7.0098612932.Signal.CY3}}{a numeric vector}
    \item{\code{ThyFC_T7.0098612932.Signal.CY5}}{a numeric vector}
    \item{\code{ThyFC_T7.0098612932.Avg_NBEADS}}{a numeric vector}
    \item{\code{ThyFC_T7.0098612932.BEAD_STDERR}}{a numeric vector}
    \item{\code{ThyFVPTC_T7.0098612931.AVG_Beta}}{a numeric vector}
    \item{\code{ThyFVPTC_T7.0098612931.Signal.CY3}}{a numeric vector}
    \item{\code{ThyFVPTC_T7.0098612931.Signal.CY5}}{a numeric vector}
    \item{\code{ThyFVPTC_T7.0098612931.Avg_NBEADS}}{a numeric vector}
    \item{\code{ThyFVPTC_T7.0098612931.BEAD_STDERR}}{a numeric vector}
    \item{\code{ThyFA_N8.0098612928.AVG_Beta}}{a numeric vector}
    \item{\code{ThyFA_N8.0098612928.Signal.CY3}}{a numeric vector}
    \item{\code{ThyFA_N8.0098612928.Signal.CY5}}{a numeric vector}
    \item{\code{ThyFA_N8.0098612928.Avg_NBEADS}}{a numeric vector}
    \item{\code{ThyFA_N8.0098612928.BEAD_STDERR}}{a numeric vector}
    \item{\code{ThyFC_T8.0098612929.AVG_Beta}}{a numeric vector}
    \item{\code{ThyFC_T8.0098612929.Signal.CY3}}{a numeric vector}
    \item{\code{ThyFC_T8.0098612929.Signal.CY5}}{a numeric vector}
    \item{\code{ThyFC_T8.0098612929.Avg_NBEADS}}{a numeric vector}
    \item{\code{ThyFC_T8.0098612929.BEAD_STDERR}}{a numeric vector}
    \item{\code{ThyFVPTC_T8.0098612930.AVG_Beta}}{a numeric vector}
    \item{\code{ThyFVPTC_T8.0098612930.Signal.CY3}}{a numeric vector}
    \item{\code{ThyFVPTC_T8.0098612930.Signal.CY5}}{a numeric vector}
    \item{\code{ThyFVPTC_T8.0098612930.Avg_NBEADS}}{a numeric vector}
    \item{\code{ThyFVPTC_T8.0098612930.BEAD_STDERR}}{a numeric vector}
    \item{\code{ThyFC_T9.0098612909.AVG_Beta}}{a numeric vector}
    \item{\code{ThyFC_T9.0098612909.Signal.CY3}}{a numeric vector}
    \item{\code{ThyFC_T9.0098612909.Signal.CY5}}{a numeric vector}
    \item{\code{ThyFC_T9.0098612909.Avg_NBEADS}}{a numeric vector}
    \item{\code{ThyFC_T9.0098612909.BEAD_STDERR}}{a numeric vector}
    \item{\code{ThyFVPTC_T9.0098612908.AVG_Beta}}{a numeric vector}
    \item{\code{ThyFVPTC_T9.0098612908.Signal.CY3}}{a numeric vector}
    \item{\code{ThyFVPTC_T9.0098612908.Signal.CY5}}{a numeric vector}
    \item{\code{ThyFVPTC_T9.0098612908.Avg_NBEADS}}{a numeric vector}
    \item{\code{ThyFVPTC_T9.0098612908.BEAD_STDERR}}{a numeric vector}
    \item{\code{ThyFVPTC_T10.0098612904.AVG_Beta}}{a numeric vector}
    \item{\code{ThyFVPTC_T10.0098612904.Signal.CY3}}{a numeric vector}
    \item{\code{ThyFVPTC_T10.0098612904.Signal.CY5}}{a numeric vector}
    \item{\code{ThyFVPTC_T10.0098612904.Avg_NBEADS}}{a numeric vector}
    \item{\code{ThyFVPTC_T10.0098612904.BEAD_STDERR}}{a numeric vector}
    \item{\code{Adenoma_T18.0098612886.AVG_Beta}}{a numeric vector}
    \item{\code{Adenoma_T18.0098612886.Signal.CY3}}{a numeric vector}
    \item{\code{Adenoma_T18.0098612886.Signal.CY5}}{a numeric vector}
    \item{\code{Adenoma_T18.0098612886.Avg_NBEADS}}{a numeric vector}
    \item{\code{Adenoma_T18.0098612886.BEAD_STDERR}}{a numeric vector}
    \item{\code{Breast_T27.0098612885.AVG_Beta}}{a numeric vector}
    \item{\code{Breast_T27.0098612885.Signal.CY3}}{a numeric vector}
    \item{\code{Breast_T27.0098612885.Signal.CY5}}{a numeric vector}
    \item{\code{Breast_T27.0098612885.Avg_NBEADS}}{a numeric vector}
    \item{\code{Breast_T27.0098612885.BEAD_STDERR}}{a numeric vector}
    \item{\code{Breast_N27.0098612884.AVG_Beta}}{a numeric vector}
    \item{\code{Breast_N27.0098612884.Signal.CY3}}{a numeric vector}
    \item{\code{Breast_N27.0098612884.Signal.CY5}}{a numeric vector}
    \item{\code{Breast_N27.0098612884.Avg_NBEADS}}{a numeric vector}
    \item{\code{Breast_N27.0098612884.BEAD_STDERR}}{a numeric vector}
    \item{\code{Breast_T28.0098612883.AVG_Beta}}{a numeric vector}
    \item{\code{Breast_T28.0098612883.Signal.CY3}}{a numeric vector}
    \item{\code{Breast_T28.0098612883.Signal.CY5}}{a numeric vector}
    \item{\code{Breast_T28.0098612883.Avg_NBEADS}}{a numeric vector}
    \item{\code{Breast_T28.0098612883.BEAD_STDERR}}{a numeric vector}
    \item{\code{Breast_N31.0098609762.AVG_Beta}}{a numeric vector}
    \item{\code{Breast_N31.0098609762.Signal.CY3}}{a numeric vector}
    \item{\code{Breast_N31.0098609762.Signal.CY5}}{a numeric vector}
    \item{\code{Breast_N31.0098609762.Avg_NBEADS}}{a numeric vector}
    \item{\code{Breast_N31.0098609762.BEAD_STDERR}}{a numeric vector}
    \item{\code{Colon_N32.0098612881.AVG_Beta}}{a numeric vector}
    \item{\code{Colon_N32.0098612881.Signal.CY3}}{a numeric vector}
    \item{\code{Colon_N32.0098612881.Signal.CY5}}{a numeric vector}
    \item{\code{Colon_N32.0098612881.Avg_NBEADS}}{a numeric vector}
    \item{\code{Colon_N32.0098612881.BEAD_STDERR}}{a numeric vector}
    \item{\code{Colon_T32.0098612882.AVG_Beta}}{a numeric vector}
    \item{\code{Colon_T32.0098612882.Signal.CY3}}{a numeric vector}
    \item{\code{Colon_T32.0098612882.Signal.CY5}}{a numeric vector}
    \item{\code{Colon_T32.0098612882.Avg_NBEADS}}{a numeric vector}
    \item{\code{Colon_T32.0098612882.BEAD_STDERR}}{a numeric vector}
    \item{\code{Adenoma_T1.0098609735.AVG_Beta}}{a numeric vector}
    \item{\code{Adenoma_T1.0098609735.Signal.CY3}}{a numeric vector}
    \item{\code{Adenoma_T1.0098609735.Signal.CY5}}{a numeric vector}
    \item{\code{Adenoma_T1.0098609735.Avg_NBEADS}}{a numeric vector}
    \item{\code{Adenoma_T1.0098609735.BEAD_STDERR}}{a numeric vector}
    \item{\code{DCIS_N1.0098611332.AVG_Beta}}{a numeric vector}
    \item{\code{DCIS_N1.0098611332.Signal.CY3}}{a numeric vector}
    \item{\code{DCIS_N1.0098611332.Signal.CY5}}{a numeric vector}
    \item{\code{DCIS_N1.0098611332.Avg_NBEADS}}{a numeric vector}
    \item{\code{DCIS_N1.0098611332.BEAD_STDERR}}{a numeric vector}
    \item{\code{ThyPTC_N1.0098609733.AVG_Beta}}{a numeric vector}
    \item{\code{ThyPTC_N1.0098609733.Signal.CY3}}{a numeric vector}
    \item{\code{ThyPTC_N1.0098609733.Signal.CY5}}{a numeric vector}
    \item{\code{ThyPTC_N1.0098609733.Avg_NBEADS}}{a numeric vector}
    \item{\code{ThyPTC_N1.0098609733.BEAD_STDERR}}{a numeric vector}
    \item{\code{Adenoma_T2.0098611310.AVG_Beta}}{a numeric vector}
    \item{\code{Adenoma_T2.0098611310.Signal.CY3}}{a numeric vector}
    \item{\code{Adenoma_T2.0098611310.Signal.CY5}}{a numeric vector}
    \item{\code{Adenoma_T2.0098611310.Avg_NBEADS}}{a numeric vector}
    \item{\code{Adenoma_T2.0098611310.BEAD_STDERR}}{a numeric vector}
    \item{\code{DCIS_N2.0098609741.AVG_Beta}}{a numeric vector}
    \item{\code{DCIS_N2.0098609741.Signal.CY3}}{a numeric vector}
    \item{\code{DCIS_N2.0098609741.Signal.CY5}}{a numeric vector}
    \item{\code{DCIS_N2.0098609741.Avg_NBEADS}}{a numeric vector}
    \item{\code{DCIS_N2.0098609741.BEAD_STDERR}}{a numeric vector}
    \item{\code{ThyPTC_T2.0098609731.AVG_Beta}}{a numeric vector}
    \item{\code{ThyPTC_T2.0098609731.Signal.CY3}}{a numeric vector}
    \item{\code{ThyPTC_T2.0098609731.Signal.CY5}}{a numeric vector}
    \item{\code{ThyPTC_T2.0098609731.Avg_NBEADS}}{a numeric vector}
    \item{\code{ThyPTC_T2.0098609731.BEAD_STDERR}}{a numeric vector}
    \item{\code{ThyPTC_N2.0098611288.AVG_Beta}}{a numeric vector}
    \item{\code{ThyPTC_N2.0098611288.Signal.CY3}}{a numeric vector}
    \item{\code{ThyPTC_N2.0098611288.Signal.CY5}}{a numeric vector}
    \item{\code{ThyPTC_N2.0098611288.Avg_NBEADS}}{a numeric vector}
    \item{\code{ThyPTC_N2.0098611288.BEAD_STDERR}}{a numeric vector}
    \item{\code{DCIS_N3.0098611285.AVG_Beta}}{a numeric vector}
    \item{\code{DCIS_N3.0098611285.Signal.CY3}}{a numeric vector}
    \item{\code{DCIS_N3.0098611285.Signal.CY5}}{a numeric vector}
    \item{\code{DCIS_N3.0098611285.Avg_NBEADS}}{a numeric vector}
    \item{\code{DCIS_N3.0098611285.BEAD_STDERR}}{a numeric vector}
    \item{\code{ThyPTC_N3.0098611263.AVG_Beta}}{a numeric vector}
    \item{\code{ThyPTC_N3.0098611263.Signal.CY3}}{a numeric vector}
    \item{\code{ThyPTC_N3.0098611263.Signal.CY5}}{a numeric vector}
    \item{\code{ThyPTC_N3.0098611263.Avg_NBEADS}}{a numeric vector}
    \item{\code{ThyPTC_N3.0098611263.BEAD_STDERR}}{a numeric vector}
    \item{\code{ThyFA_N4.0098611260.AVG_Beta}}{a numeric vector}
    \item{\code{ThyFA_N4.0098611260.Signal.CY3}}{a numeric vector}
    \item{\code{ThyFA_N4.0098611260.Signal.CY5}}{a numeric vector}
    \item{\code{ThyFA_N4.0098611260.Avg_NBEADS}}{a numeric vector}
    \item{\code{ThyFA_N4.0098611260.BEAD_STDERR}}{a numeric vector}
    \item{\code{ThyFC_T4.0098611259.AVG_Beta}}{a numeric vector}
    \item{\code{ThyFC_T4.0098611259.Signal.CY3}}{a numeric vector}
    \item{\code{ThyFC_T4.0098611259.Signal.CY5}}{a numeric vector}
    \item{\code{ThyFC_T4.0098611259.Avg_NBEADS}}{a numeric vector}
    \item{\code{ThyFC_T4.0098611259.BEAD_STDERR}}{a numeric vector}
    \item{\code{ThyFA_N1.0098611330.AVG_Beta}}{a numeric vector}
    \item{\code{ThyFA_N1.0098611330.Signal.CY3}}{a numeric vector}
    \item{\code{ThyFA_N1.0098611330.Signal.CY5}}{a numeric vector}
    \item{\code{ThyFA_N1.0098611330.Avg_NBEADS}}{a numeric vector}
    \item{\code{ThyFA_N1.0098611330.BEAD_STDERR}}{a numeric vector}
    \item{\code{ThyFC_T1.0098611328.AVG_Beta}}{a numeric vector}
    \item{\code{ThyFC_T1.0098611328.Signal.CY3}}{a numeric vector}
    \item{\code{ThyFC_T1.0098611328.Signal.CY5}}{a numeric vector}
    \item{\code{ThyFC_T1.0098611328.Avg_NBEADS}}{a numeric vector}
    \item{\code{ThyFC_T1.0098611328.BEAD_STDERR}}{a numeric vector}
    \item{\code{ThyFVPTC_T1.0098611327.AVG_Beta}}{a numeric vector}
    \item{\code{ThyFVPTC_T1.0098611327.Signal.CY3}}{a numeric vector}
    \item{\code{ThyFVPTC_T1.0098611327.Signal.CY5}}{a numeric vector}
    \item{\code{ThyFVPTC_T1.0098611327.Avg_NBEADS}}{a numeric vector}
    \item{\code{ThyFVPTC_T1.0098611327.BEAD_STDERR}}{a numeric vector}
    \item{\code{ThyFA_N2.0098611305.AVG_Beta}}{a numeric vector}
    \item{\code{ThyFA_N2.0098611305.Signal.CY3}}{a numeric vector}
    \item{\code{ThyFA_N2.0098611305.Signal.CY5}}{a numeric vector}
    \item{\code{ThyFA_N2.0098611305.Avg_NBEADS}}{a numeric vector}
    \item{\code{ThyFA_N2.0098611305.BEAD_STDERR}}{a numeric vector}
    \item{\code{ThyFC_T2.0098611304.AVG_Beta}}{a numeric vector}
    \item{\code{ThyFC_T2.0098611304.Signal.CY3}}{a numeric vector}
    \item{\code{ThyFC_T2.0098611304.Signal.CY5}}{a numeric vector}
    \item{\code{ThyFC_T2.0098611304.Avg_NBEADS}}{a numeric vector}
    \item{\code{ThyFC_T2.0098611304.BEAD_STDERR}}{a numeric vector}
    \item{\code{ThyFVPTC_T2.0098611303.AVG_Beta}}{a numeric vector}
    \item{\code{ThyFVPTC_T2.0098611303.Signal.CY3}}{a numeric vector}
    \item{\code{ThyFVPTC_T2.0098611303.Signal.CY5}}{a numeric vector}
    \item{\code{ThyFVPTC_T2.0098611303.Avg_NBEADS}}{a numeric vector}
    \item{\code{ThyFVPTC_T2.0098611303.BEAD_STDERR}}{a numeric vector}
    \item{\code{ThyFC_T3.0098611281.AVG_Beta}}{a numeric vector}
    \item{\code{ThyFC_T3.0098611281.Signal.CY3}}{a numeric vector}
    \item{\code{ThyFC_T3.0098611281.Signal.CY5}}{a numeric vector}
    \item{\code{ThyFC_T3.0098611281.Avg_NBEADS}}{a numeric vector}
    \item{\code{ThyFC_T3.0098611281.BEAD_STDERR}}{a numeric vector}
    \item{\code{ThyFVPTC_T3.0098611279.AVG_Beta}}{a numeric vector}
    \item{\code{ThyFVPTC_T3.0098611279.Signal.CY3}}{a numeric vector}
    \item{\code{ThyFVPTC_T3.0098611279.Signal.CY5}}{a numeric vector}
    \item{\code{ThyFVPTC_T3.0098611279.Avg_NBEADS}}{a numeric vector}
    \item{\code{ThyFVPTC_T3.0098611279.BEAD_STDERR}}{a numeric vector}
    \item{\code{ThyFC_T10.0098611270.AVG_Beta}}{a numeric vector}
    \item{\code{ThyFC_T10.0098611270.Signal.CY3}}{a numeric vector}
    \item{\code{ThyFC_T10.0098611270.Signal.CY5}}{a numeric vector}
    \item{\code{ThyFC_T10.0098611270.Avg_NBEADS}}{a numeric vector}
    \item{\code{ThyFC_T10.0098611270.BEAD_STDERR}}{a numeric vector}
    \item{\code{ThyFVPTC_T4.0098611258.AVG_Beta}}{a numeric vector}
    \item{\code{ThyFVPTC_T4.0098611258.Signal.CY3}}{a numeric vector}
    \item{\code{ThyFVPTC_T4.0098611258.Signal.CY5}}{a numeric vector}
    \item{\code{ThyFVPTC_T4.0098611258.Avg_NBEADS}}{a numeric vector}
    \item{\code{ThyFVPTC_T4.0098611258.BEAD_STDERR}}{a numeric vector}
    \item{\code{ThyFVPTC_N4.0098609745.AVG_Beta}}{a numeric vector}
    \item{\code{ThyFVPTC_N4.0098609745.Signal.CY3}}{a numeric vector}
    \item{\code{ThyFVPTC_N4.0098609745.Signal.CY5}}{a numeric vector}
    \item{\code{ThyFVPTC_N4.0098609745.Avg_NBEADS}}{a numeric vector}
    \item{\code{ThyFVPTC_N4.0098609745.BEAD_STDERR}}{a numeric vector}
    \item{\code{Adenoma_T5.0098611244.AVG_Beta}}{a numeric vector}
    \item{\code{Adenoma_T5.0098611244.Signal.CY3}}{a numeric vector}
    \item{\code{Adenoma_T5.0098611244.Signal.CY5}}{a numeric vector}
    \item{\code{Adenoma_T5.0098611244.Avg_NBEADS}}{a numeric vector}
    \item{\code{Adenoma_T5.0098611244.BEAD_STDERR}}{a numeric vector}
    \item{\code{ThyFC_T5.0098612966.AVG_Beta}}{a numeric vector}
    \item{\code{ThyFC_T5.0098612966.Signal.CY3}}{a numeric vector}
    \item{\code{ThyFC_T5.0098612966.Signal.CY5}}{a numeric vector}
    \item{\code{ThyFC_T5.0098612966.Avg_NBEADS}}{a numeric vector}
    \item{\code{ThyFC_T5.0098612966.BEAD_STDERR}}{a numeric vector}
    \item{\code{ThyFVPTC_T5.0098612964.AVG_Beta}}{a numeric vector}
    \item{\code{ThyFVPTC_T5.0098612964.Signal.CY3}}{a numeric vector}
    \item{\code{ThyFVPTC_T5.0098612964.Signal.CY5}}{a numeric vector}
    \item{\code{ThyFVPTC_T5.0098612964.Avg_NBEADS}}{a numeric vector}
    \item{\code{ThyFVPTC_T5.0098612964.BEAD_STDERR}}{a numeric vector}
    \item{\code{ThyFVPTC_N5.0098612963.AVG_Beta}}{a numeric vector}
    \item{\code{ThyFVPTC_N5.0098612963.Signal.CY3}}{a numeric vector}
    \item{\code{ThyFVPTC_N5.0098612963.Signal.CY5}}{a numeric vector}
    \item{\code{ThyFVPTC_N5.0098612963.Avg_NBEADS}}{a numeric vector}
    \item{\code{ThyFVPTC_N5.0098612963.BEAD_STDERR}}{a numeric vector}
    \item{\code{Adenoma_T6.0098612943.AVG_Beta}}{a numeric vector}
    \item{\code{Adenoma_T6.0098612943.Signal.CY3}}{a numeric vector}
    \item{\code{Adenoma_T6.0098612943.Signal.CY5}}{a numeric vector}
    \item{\code{Adenoma_T6.0098612943.Avg_NBEADS}}{a numeric vector}
    \item{\code{Adenoma_T6.0098612943.BEAD_STDERR}}{a numeric vector}
    \item{\code{ThyFVPTC_N7.0098612919.AVG_Beta}}{a numeric vector}
    \item{\code{ThyFVPTC_N7.0098612919.Signal.CY3}}{a numeric vector}
    \item{\code{ThyFVPTC_N7.0098612919.Signal.CY5}}{a numeric vector}
    \item{\code{ThyFVPTC_N7.0098612919.Avg_NBEADS}}{a numeric vector}
    \item{\code{ThyFVPTC_N7.0098612919.BEAD_STDERR}}{a numeric vector}
    \item{\code{ThyHC_T7.0098612921.AVG_Beta}}{a numeric vector}
    \item{\code{ThyHC_T7.0098612921.Signal.CY3}}{a numeric vector}
    \item{\code{ThyHC_T7.0098612921.Signal.CY5}}{a numeric vector}
    \item{\code{ThyHC_T7.0098612921.Avg_NBEADS}}{a numeric vector}
    \item{\code{ThyHC_T7.0098612921.BEAD_STDERR}}{a numeric vector}
    \item{\code{ThyPTC_T7.0098612922.AVG_Beta}}{a numeric vector}
    \item{\code{ThyPTC_T7.0098612922.Signal.CY3}}{a numeric vector}
    \item{\code{ThyPTC_T7.0098612922.Signal.CY5}}{a numeric vector}
    \item{\code{ThyPTC_T7.0098612922.Avg_NBEADS}}{a numeric vector}
    \item{\code{ThyPTC_T7.0098612922.BEAD_STDERR}}{a numeric vector}
    \item{\code{ThyHC_T8.0098612917.AVG_Beta}}{a numeric vector}
    \item{\code{ThyHC_T8.0098612917.Signal.CY3}}{a numeric vector}
    \item{\code{ThyHC_T8.0098612917.Signal.CY5}}{a numeric vector}
    \item{\code{ThyHC_T8.0098612917.Avg_NBEADS}}{a numeric vector}
    \item{\code{ThyHC_T8.0098612917.BEAD_STDERR}}{a numeric vector}
    \item{\code{ThyPTC_T8.0098612916.AVG_Beta}}{a numeric vector}
    \item{\code{ThyPTC_T8.0098612916.Signal.CY3}}{a numeric vector}
    \item{\code{ThyPTC_T8.0098612916.Signal.CY5}}{a numeric vector}
    \item{\code{ThyPTC_T8.0098612916.Avg_NBEADS}}{a numeric vector}
    \item{\code{ThyPTC_T8.0098612916.BEAD_STDERR}}{a numeric vector}
    \item{\code{ThyPTC_N8.0098612915.AVG_Beta}}{a numeric vector}
    \item{\code{ThyPTC_N8.0098612915.Signal.CY3}}{a numeric vector}
    \item{\code{ThyPTC_N8.0098612915.Signal.CY5}}{a numeric vector}
    \item{\code{ThyPTC_N8.0098612915.Avg_NBEADS}}{a numeric vector}
    \item{\code{ThyPTC_N8.0098612915.BEAD_STDERR}}{a numeric vector}
    \item{\code{ThyPTC_T9.0098612896.AVG_Beta}}{a numeric vector}
    \item{\code{ThyPTC_T9.0098612896.Signal.CY3}}{a numeric vector}
    \item{\code{ThyPTC_T9.0098612896.Signal.CY5}}{a numeric vector}
    \item{\code{ThyPTC_T9.0098612896.Avg_NBEADS}}{a numeric vector}
    \item{\code{ThyPTC_T9.0098612896.BEAD_STDERR}}{a numeric vector}
    \item{\code{ThyPTC_N9.0098612897.AVG_Beta}}{a numeric vector}
    \item{\code{ThyPTC_N9.0098612897.Signal.CY3}}{a numeric vector}
    \item{\code{ThyPTC_N9.0098612897.Signal.CY5}}{a numeric vector}
    \item{\code{ThyPTC_N9.0098612897.Avg_NBEADS}}{a numeric vector}
    \item{\code{ThyPTC_N9.0098612897.BEAD_STDERR}}{a numeric vector}
    \item{\code{Adenoma_T10.0098612898.AVG_Beta}}{a numeric vector}
    \item{\code{Adenoma_T10.0098612898.Signal.CY3}}{a numeric vector}
    \item{\code{Adenoma_T10.0098612898.Signal.CY5}}{a numeric vector}
    \item{\code{Adenoma_T10.0098612898.Avg_NBEADS}}{a numeric vector}
    \item{\code{Adenoma_T10.0098612898.BEAD_STDERR}}{a numeric vector}
    \item{\code{ThyPTC_N10.0098612894.AVG_Beta}}{a numeric vector}
    \item{\code{ThyPTC_N10.0098612894.Signal.CY3}}{a numeric vector}
    \item{\code{ThyPTC_N10.0098612894.Signal.CY5}}{a numeric vector}
    \item{\code{ThyPTC_N10.0098612894.Avg_NBEADS}}{a numeric vector}
    \item{\code{ThyPTC_N10.0098612894.BEAD_STDERR}}{a numeric vector}
    \item{\code{Adenoma_T11.0098612893.AVG_Beta}}{a numeric vector}
    \item{\code{Adenoma_T11.0098612893.Signal.CY3}}{a numeric vector}
    \item{\code{Adenoma_T11.0098612893.Signal.CY5}}{a numeric vector}
    \item{\code{Adenoma_T11.0098612893.Avg_NBEADS}}{a numeric vector}
    \item{\code{Adenoma_T11.0098612893.BEAD_STDERR}}{a numeric vector}
    \item{\code{Adenoma_T12.0098612892.AVG_Beta}}{a numeric vector}
    \item{\code{Adenoma_T12.0098612892.Signal.CY3}}{a numeric vector}
    \item{\code{Adenoma_T12.0098612892.Signal.CY5}}{a numeric vector}
    \item{\code{Adenoma_T12.0098612892.Avg_NBEADS}}{a numeric vector}
    \item{\code{Adenoma_T12.0098612892.BEAD_STDERR}}{a numeric vector}
    \item{\code{Adenoma_T13.0098612891.AVG_Beta}}{a numeric vector}
    \item{\code{Adenoma_T13.0098612891.Signal.CY3}}{a numeric vector}
    \item{\code{Adenoma_T13.0098612891.Signal.CY5}}{a numeric vector}
    \item{\code{Adenoma_T13.0098612891.Avg_NBEADS}}{a numeric vector}
    \item{\code{Adenoma_T13.0098612891.BEAD_STDERR}}{a numeric vector}
    \item{\code{Breast_N28.0098612871.AVG_Beta}}{a numeric vector}
    \item{\code{Breast_N28.0098612871.Signal.CY3}}{a numeric vector}
    \item{\code{Breast_N28.0098612871.Signal.CY5}}{a numeric vector}
    \item{\code{Breast_N28.0098612871.Avg_NBEADS}}{a numeric vector}
    \item{\code{Breast_N28.0098612871.BEAD_STDERR}}{a numeric vector}
    \item{\code{Breast_N29.0098612873.AVG_Beta}}{a numeric vector}
    \item{\code{Breast_N29.0098612873.Signal.CY3}}{a numeric vector}
    \item{\code{Breast_N29.0098612873.Signal.CY5}}{a numeric vector}
    \item{\code{Breast_N29.0098612873.Avg_NBEADS}}{a numeric vector}
    \item{\code{Breast_N29.0098612873.BEAD_STDERR}}{a numeric vector}
    \item{\code{Breast_T30.0098612874.AVG_Beta}}{a numeric vector}
    \item{\code{Breast_T30.0098612874.Signal.CY3}}{a numeric vector}
    \item{\code{Breast_T30.0098612874.Signal.CY5}}{a numeric vector}
    \item{\code{Breast_T30.0098612874.Avg_NBEADS}}{a numeric vector}
    \item{\code{Breast_T30.0098612874.BEAD_STDERR}}{a numeric vector}
    \item{\code{ThyFVPTC_N1.0098611326.AVG_Beta}}{a numeric vector}
    \item{\code{ThyFVPTC_N1.0098611326.Signal.CY3}}{a numeric vector}
    \item{\code{ThyFVPTC_N1.0098611326.Signal.CY5}}{a numeric vector}
    \item{\code{ThyFVPTC_N1.0098611326.Avg_NBEADS}}{a numeric vector}
    \item{\code{ThyFVPTC_N1.0098611326.BEAD_STDERR}}{a numeric vector}
    \item{\code{ThyHC_T1.0098611324.AVG_Beta}}{a numeric vector}
    \item{\code{ThyHC_T1.0098611324.Signal.CY3}}{a numeric vector}
    \item{\code{ThyHC_T1.0098611324.Signal.CY5}}{a numeric vector}
    \item{\code{ThyHC_T1.0098611324.Avg_NBEADS}}{a numeric vector}
    \item{\code{ThyHC_T1.0098611324.BEAD_STDERR}}{a numeric vector}
    \item{\code{ThyPTC_T1.0098611323.AVG_Beta}}{a numeric vector}
    \item{\code{ThyPTC_T1.0098611323.Signal.CY3}}{a numeric vector}
    \item{\code{ThyPTC_T1.0098611323.Signal.CY5}}{a numeric vector}
    \item{\code{ThyPTC_T1.0098611323.Avg_NBEADS}}{a numeric vector}
    \item{\code{ThyPTC_T1.0098611323.BEAD_STDERR}}{a numeric vector}
    \item{\code{ThyHC_T6.0098611320.AVG_Beta}}{a numeric vector}
    \item{\code{ThyHC_T6.0098611320.Signal.CY3}}{a numeric vector}
    \item{\code{ThyHC_T6.0098611320.Signal.CY5}}{a numeric vector}
    \item{\code{ThyHC_T6.0098611320.Avg_NBEADS}}{a numeric vector}
    \item{\code{ThyHC_T6.0098611320.BEAD_STDERR}}{a numeric vector}
    \item{\code{ThyFVPTC_N2.0098611302.AVG_Beta}}{a numeric vector}
    \item{\code{ThyFVPTC_N2.0098611302.Signal.CY3}}{a numeric vector}
    \item{\code{ThyFVPTC_N2.0098611302.Signal.CY5}}{a numeric vector}
    \item{\code{ThyFVPTC_N2.0098611302.Avg_NBEADS}}{a numeric vector}
    \item{\code{ThyFVPTC_N2.0098611302.BEAD_STDERR}}{a numeric vector}
    \item{\code{ThyHC_T2.0098611299.AVG_Beta}}{a numeric vector}
    \item{\code{ThyHC_T2.0098611299.Signal.CY3}}{a numeric vector}
    \item{\code{ThyHC_T2.0098611299.Signal.CY5}}{a numeric vector}
    \item{\code{ThyHC_T2.0098611299.Avg_NBEADS}}{a numeric vector}
    \item{\code{ThyHC_T2.0098611299.BEAD_STDERR}}{a numeric vector}
    \item{\code{Adenoma_T3.0098611298.AVG_Beta}}{a numeric vector}
    \item{\code{Adenoma_T3.0098611298.Signal.CY3}}{a numeric vector}
    \item{\code{Adenoma_T3.0098611298.Signal.CY5}}{a numeric vector}
    \item{\code{Adenoma_T3.0098611298.Avg_NBEADS}}{a numeric vector}
    \item{\code{Adenoma_T3.0098611298.BEAD_STDERR}}{a numeric vector}
    \item{\code{ThyFVPTC_N3.0098609742.AVG_Beta}}{a numeric vector}
    \item{\code{ThyFVPTC_N3.0098609742.Signal.CY3}}{a numeric vector}
    \item{\code{ThyFVPTC_N3.0098609742.Signal.CY5}}{a numeric vector}
    \item{\code{ThyFVPTC_N3.0098609742.Avg_NBEADS}}{a numeric vector}
    \item{\code{ThyFVPTC_N3.0098609742.BEAD_STDERR}}{a numeric vector}
    \item{\code{ThyHC_T3.0098611276.AVG_Beta}}{a numeric vector}
    \item{\code{ThyHC_T3.0098611276.Signal.CY3}}{a numeric vector}
    \item{\code{ThyHC_T3.0098611276.Signal.CY5}}{a numeric vector}
    \item{\code{ThyHC_T3.0098611276.Avg_NBEADS}}{a numeric vector}
    \item{\code{ThyHC_T3.0098611276.BEAD_STDERR}}{a numeric vector}
    \item{\code{ThyPTC_T3.0098611275.AVG_Beta}}{a numeric vector}
    \item{\code{ThyPTC_T3.0098611275.Signal.CY3}}{a numeric vector}
    \item{\code{ThyPTC_T3.0098611275.Signal.CY5}}{a numeric vector}
    \item{\code{ThyPTC_T3.0098611275.Avg_NBEADS}}{a numeric vector}
    \item{\code{ThyPTC_T3.0098611275.BEAD_STDERR}}{a numeric vector}
    \item{\code{Adenoma_T4.0098611272.AVG_Beta}}{a numeric vector}
    \item{\code{Adenoma_T4.0098611272.Signal.CY3}}{a numeric vector}
    \item{\code{Adenoma_T4.0098611272.Signal.CY5}}{a numeric vector}
    \item{\code{Adenoma_T4.0098611272.Avg_NBEADS}}{a numeric vector}
    \item{\code{Adenoma_T4.0098611272.BEAD_STDERR}}{a numeric vector}
    \item{\code{DCIS_N4.0098611274.AVG_Beta}}{a numeric vector}
    \item{\code{DCIS_N4.0098611274.Signal.CY3}}{a numeric vector}
    \item{\code{DCIS_N4.0098611274.Signal.CY5}}{a numeric vector}
    \item{\code{DCIS_N4.0098611274.Avg_NBEADS}}{a numeric vector}
    \item{\code{DCIS_N4.0098611274.BEAD_STDERR}}{a numeric vector}
    \item{\code{ThyHC_T4.0098611254.AVG_Beta}}{a numeric vector}
    \item{\code{ThyHC_T4.0098611254.Signal.CY3}}{a numeric vector}
    \item{\code{ThyHC_T4.0098611254.Signal.CY5}}{a numeric vector}
    \item{\code{ThyHC_T4.0098611254.Avg_NBEADS}}{a numeric vector}
    \item{\code{ThyHC_T4.0098611254.BEAD_STDERR}}{a numeric vector}
    \item{\code{ThyPTC_T4.0098611253.AVG_Beta}}{a numeric vector}
    \item{\code{ThyPTC_T4.0098611253.Signal.CY3}}{a numeric vector}
    \item{\code{ThyPTC_T4.0098611253.Signal.CY5}}{a numeric vector}
    \item{\code{ThyPTC_T4.0098611253.Avg_NBEADS}}{a numeric vector}
    \item{\code{ThyPTC_T4.0098611253.BEAD_STDERR}}{a numeric vector}
    \item{\code{ThyPTC_N4.0098611252.AVG_Beta}}{a numeric vector}
    \item{\code{ThyPTC_N4.0098611252.Signal.CY3}}{a numeric vector}
    \item{\code{ThyPTC_N4.0098611252.Signal.CY5}}{a numeric vector}
    \item{\code{ThyPTC_N4.0098611252.Avg_NBEADS}}{a numeric vector}
    \item{\code{ThyPTC_N4.0098611252.BEAD_STDERR}}{a numeric vector}
    \item{\code{DCIS_N5.0098611247.AVG_Beta}}{a numeric vector}
    \item{\code{DCIS_N5.0098611247.Signal.CY3}}{a numeric vector}
    \item{\code{DCIS_N5.0098611247.Signal.CY5}}{a numeric vector}
    \item{\code{DCIS_N5.0098611247.Avg_NBEADS}}{a numeric vector}
    \item{\code{DCIS_N5.0098611247.BEAD_STDERR}}{a numeric vector}
    \item{\code{ThyFA_N5.0098609744.AVG_Beta}}{a numeric vector}
    \item{\code{ThyFA_N5.0098609744.Signal.CY3}}{a numeric vector}
    \item{\code{ThyFA_N5.0098609744.Signal.CY5}}{a numeric vector}
    \item{\code{ThyFA_N5.0098609744.Avg_NBEADS}}{a numeric vector}
    \item{\code{ThyFA_N5.0098609744.BEAD_STDERR}}{a numeric vector}
  }
}
\references{
Hansen KD, et al. (2011) Increased methylation variation in epigenetic domains 
across cancer types. Nat Genet 43(8):768-775
}
\examples{
data(Methylation)
}
\keyword{datasets}
